@@ -1,7 +1,7 @@
-from typing import TYPE_CHECKING, List, Tuple, Union
+from typing import TYPE_CHECKING, Union
 
 if TYPE_CHECKING:
     from .vec2d import Vec2d
 
-_Vec2dOrTuple = Union[Tuple[float, float], List[float], "Vec2d"]
-_Vec2dOrFloat = Union[Tuple[float, float], List[float], "Vec2d", float]
+_Vec2dOrTuple = Union[tuple[float, float], list[float], "Vec2d"]
+_Vec2dOrFloat = Union[tuple[float, float], list[float], "Vec2d", float]
