@@ -1,16 +1,10 @@
 __docformat__ = "reStructuredText"
 
-from typing import (  # Literal,
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    ClassVar,
-    Optional,
-    Set,
-    Tuple,
-    Union,
-)
-from weakref import WeakSet
+import math
+import weakref
+from collections.abc import KeysView
+from typing import TYPE_CHECKING, Any, Callable, ClassVar, Optional  # Literal,
+from weakref import WeakKeyDictionary
 
 if TYPE_CHECKING:
     from .space import Space
@@ -20,6 +14,8 @@ if TYPE_CHECKING:
 from ._chipmunk_cffi import ffi, lib
 from ._pickle import PickleMixin, _State
 from ._typing_attr import TypingAttrMixing
+from ._util import _dead_ref
+from ._weakkeysview import WeakKeysView
 from .vec2d import Vec2d
 
 _BodyType = int
@@ -106,12 +102,11 @@ class Body(PickleMixin, TypingAttrMixing, object):
         "is_sleeping",
         "_velocity_func",
         "_position_func",
+        # "_space",
     ]
 
     _position_func: Optional[_PositionFunc] = None
     _velocity_func: Optional[_VelocityFunc] = None
-
-    _id_counter = 1
 
     def __init__(
         self, mass: float = 0, moment: float = 0, body_type: _BodyType = DYNAMIC
@@ -217,14 +212,9 @@ class Body(PickleMixin, TypingAttrMixing, object):
         elif body_type == Body.STATIC:
             self._body = ffi.gc(lib.cpBodyNewStatic(), freebody)
 
-        self._space: Optional["Space"] = (
-            None  # Weak ref to the space holding this body (if any)
-        )
-
-        self._constraints: WeakSet["Constraint"] = (
-            WeakSet()
-        )  # weak refs to any constraints attached
-        self._shapes: WeakSet["Shape"] = WeakSet()  # weak refs to any shapes attached
+        self._space: weakref.ref["Space"] = _dead_ref
+        self._constraints: WeakKeyDictionary["Constraint", None] = WeakKeyDictionary()
+        self._shapes: dict["Shape", None] = {}
 
         d = ffi.new_handle(self)
         self._data_handle = d  # to prevent gc to collect the handle
@@ -252,30 +242,40 @@ class Body(PickleMixin, TypingAttrMixing, object):
         else:
             return "Body(Body.STATIC)"
 
-    def _set_mass(self, mass: float) -> None:
-        lib.cpBodySetMass(self._body, mass)
+    @property
+    def mass(self) -> float:
+        """Mass of the body.
 
-    def _get_mass(self) -> float:
+        Note that dynamic bodies must have mass > 0 if they are attached to a
+        Space.
+        """
         return lib.cpBodyGetMass(self._body)
 
-    mass = property(_get_mass, _set_mass, doc="""Mass of the body.""")
+    @mass.setter
+    def mass(self, mass: float) -> None:
+        assert (
+            self.space is None or 0 < mass < math.inf
+        ), "Dynamic bodies must have mass > 0 if they are attached to a Space."
+        lib.cpBodySetMass(self._body, mass)
 
-    def _set_moment(self, moment: float) -> None:
-        lib.cpBodySetMoment(self._body, moment)
+    @property
+    def moment(self) -> float:
+        """Moment of inertia (MoI or sometimes just moment) of the body.
 
-    def _get_moment(self) -> float:
+        The moment is like the rotational mass of a body. Note that it is
+        valid to set moment to float('inf'). This will make a body that cannot
+        rotate.
+        """
         return lib.cpBodyGetMoment(self._body)
 
-    moment = property(
-        _get_moment,
-        _set_moment,
-        doc="""Moment of inertia (MoI or sometimes just moment) of the body.
+    @moment.setter
+    def moment(self, moment: float) -> None:
+        assert (
+            self.space is None or moment > 0
+        ), "Dynamic bodies must have moment > 0 if they are attached to a Space"
+        lib.cpBodySetMoment(self._body, moment)
 
-        The moment is like the rotational mass of a body.
-        """,
-    )
-
-    def _set_position(self, pos: Union[Vec2d, Tuple[float, float]]) -> None:
+    def _set_position(self, pos: tuple[float, float]) -> None:
         assert len(pos) == 2
         lib.cpBodySetPosition(self._body, pos)
 
@@ -294,7 +294,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         queries against the space.""",
     )
 
-    def _set_center_of_gravity(self, cog: Tuple[float, float]) -> None:
+    def _set_center_of_gravity(self, cog: tuple[float, float]) -> None:
         assert len(cog) == 2
         lib.cpBodySetCenterOfGravity(self._body, cog)
 
@@ -312,7 +312,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         """,
     )
 
-    def _set_velocity(self, vel: Tuple[float, float]) -> None:
+    def _set_velocity(self, vel: tuple[float, float]) -> None:
         assert len(vel) == 2
         lib.cpBodySetVelocity(self._body, vel)
 
@@ -326,7 +326,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         doc="""Linear velocity of the center of gravity of the body.""",
     )
 
-    def _set_force(self, f: Tuple[float, float]) -> None:
+    def _set_force(self, f: tuple[float, float]) -> None:
         assert len(f) == 2
         lib.cpBodySetForce(self._body, f)
 
@@ -344,78 +344,65 @@ class Body(PickleMixin, TypingAttrMixing, object):
         force applied manually from the apply force functions.""",
     )
 
-    def _set_angle(self, angle: float) -> None:
-        lib.cpBodySetAngle(self._body, angle)
-
-    def _get_angle(self) -> float:
-        return lib.cpBodyGetAngle(self._body)
-
-    angle = property(
-        _get_angle,
-        _set_angle,
-        doc="""Rotation of the body in radians.
+    @property
+    def angle(self) -> float:
+        """Rotation of the body in radians.
 
         When changing the rotation you may also want to call
-        :py:func:`Space.reindex_shapes_for_body` to update the collision 
-        detection information for the attached shapes if plan to make any 
-        queries against the space. A body rotates around its center of gravity, 
+        :py:func:`Space.reindex_shapes_for_body` to update the collision
+        detection information for the attached shapes if plan to make any
+        queries against the space. A body rotates around its center of gravity,
         not its position.
 
         .. Note::
             If you get small/no changes to the angle when for example a
             ball is "rolling" down a slope it might be because the Circle shape
             attached to the body or the slope shape does not have any friction
-            set.""",
-    )
+            set."""
+        return lib.cpBodyGetAngle(self._body)
 
-    def _set_angular_velocity(self, w: float) -> None:
-        lib.cpBodySetAngularVelocity(self._body, w)
+    @angle.setter
+    def angle(self, angle: float) -> None:
+        lib.cpBodySetAngle(self._body, angle)
 
-    def _get_angular_velocity(self) -> float:
+    @property
+    def angular_velocity(self) -> float:
+        """The angular velocity of the body in radians per second."""
         return lib.cpBodyGetAngularVelocity(self._body)
 
-    angular_velocity = property(
-        _get_angular_velocity,
-        _set_angular_velocity,
-        doc="""The angular velocity of the body in radians per second.""",
-    )
+    @angular_velocity.setter
+    def angular_velocity(self, w: float) -> None:
+        lib.cpBodySetAngularVelocity(self._body, w)
 
-    def _set_torque(self, t: float) -> None:
-        lib.cpBodySetTorque(self._body, t)
+    @property
+    def torque(self) -> float:
+        """The torque applied to the body.
 
-    def _get_torque(self) -> float:
+        This value is reset for every time step."""
         return lib.cpBodyGetTorque(self._body)
 
-    torque = property(
-        _get_torque,
-        _set_torque,
-        doc="""The torque applied to the body.
+    @torque.setter
+    def torque(self, t: float) -> None:
+        lib.cpBodySetTorque(self._body, t)
 
-        This value is reset for every time step.""",
-    )
-
-    def _get_rotation_vector(self) -> Vec2d:
+    @property
+    def rotation_vector(self) -> Vec2d:
+        """The rotation vector for the body."""
         v = lib.cpBodyGetRotation(self._body)
         return Vec2d(v.x, v.y)
-
-    rotation_vector = property(
-        _get_rotation_vector, doc="""The rotation vector for the body."""
-    )
 
     @property
     def space(self) -> Optional["Space"]:
         """Get the :py:class:`Space` that the body has been added to (or
         None)."""
+        # This assert is tested in test_pickle_circular_ref
         assert hasattr(self, "_space"), (
             "_space not set. This can mean there's a direct or indirect"
             " circular reference between the Body and the Space. Circular"
             " references are not supported when using pickle or copy and"
             " might crash."
         )
-        if self._space is not None:
-            return self._space._get_self()  # ugly hack because of weakref
-        else:
-            return None
+        return self._space()
 
     def _set_velocity_func(self, func: _VelocityFunc) -> None:
         if func == Body.update_velocity:
@@ -495,16 +482,11 @@ class Body(PickleMixin, TypingAttrMixing, object):
     @property
     def kinetic_energy(self) -> float:
         """Get the kinetic energy of a body."""
-        # todo: use ffi method
-        # return lib._cpBodyKineticEnergy(self._body)
-
-        vsq: float = self.velocity.dot(self.velocity)
-        wsq: float = self.angular_velocity * self.angular_velocity
-        return (vsq * self.mass if vsq else 0.0) + (wsq * self.moment if wsq else 0.0)
+        return lib.cpBodyKineticEnergy(self._body)
 
     @staticmethod
     def update_velocity(
-        body: "Body", gravity: Tuple[float, float], damping: float, dt: float
+        body: "Body", gravity: tuple[float, float], damping: float, dt: float
     ) -> None:
         """Default rigid body velocity integration function.
 
@@ -526,7 +508,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         lib.cpBodyUpdatePosition(body._body, dt)
 
     def apply_force_at_world_point(
-        self, force: Tuple[float, float], point: Tuple[float, float]
+        self, force: tuple[float, float], point: tuple[float, float]
     ) -> None:
         """Add the force force to body as if applied from the world point.
 
@@ -543,7 +525,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         lib.cpBodyApplyForceAtWorldPoint(self._body, force, point)
 
     def apply_force_at_local_point(
-        self, force: Tuple[float, float], point: Tuple[float, float] = (0, 0)
+        self, force: tuple[float, float], point: tuple[float, float] = (0, 0)
     ) -> None:
         """Add the local force force to body as if applied from the body
         local point.
@@ -553,7 +535,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         lib.cpBodyApplyForceAtLocalPoint(self._body, force, point)
 
     def apply_impulse_at_world_point(
-        self, impulse: Tuple[float, float], point: Tuple[float, float]
+        self, impulse: tuple[float, float], point: tuple[float, float]
     ) -> None:
         """Add the impulse impulse to body as if applied from the world point."""
         assert len(impulse) == 2
@@ -561,7 +543,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         lib.cpBodyApplyImpulseAtWorldPoint(self._body, impulse, point)
 
     def apply_impulse_at_local_point(
-        self, impulse: Tuple[float, float], point: Tuple[float, float] = (0, 0)
+        self, impulse: tuple[float, float], point: tuple[float, float] = (0, 0)
     ) -> None:
         """Add the local impulse impulse to body as if applied from the body
         local point.
@@ -582,7 +564,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
 
         Cannot be called from a callback.
         """
-        if self._space == None:
+        if self.space == None:
             raise Exception("Body not added to space")
         lib.cpBodySleep(self._body)
 
@@ -600,7 +582,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         to initialize levels and start stacks of objects in a pre-sleeping
         state.
         """
-        if self._space == None:
+        if self.space == None:
             raise Exception("Body not added to space")
         lib.cpBodySleepWithGroup(self._body, body._body)
 
@@ -609,28 +591,31 @@ class Body(PickleMixin, TypingAttrMixing, object):
         """Returns true if the body is sleeping."""
         return bool(lib.cpBodyIsSleeping(self._body))
 
-    def _set_type(self, body_type: _BodyType) -> None:
-        lib.cpBodySetType(self._body, body_type)
-
-    def _get_type(self) -> _BodyType:
-        return lib.cpBodyGetType(self._body)
-
-    body_type = property(
-        _get_type,
-        _set_type,
-        doc="""The type of a body (:py:const:`Body.DYNAMIC`, 
+    @property
+    def body_type(self) -> _BodyType:
+        """The type of a body (:py:const:`Body.DYNAMIC`,
         :py:const:`Body.KINEMATIC` or :py:const:`Body.STATIC`).
 
         When changing an body to a dynamic body, the mass and moment of
         inertia are recalculated from the shapes added to the body. Custom
         calculated moments of inertia are not preserved when changing types.
         This function cannot be called directly in a collision callback.
-        """,
-    )
+        """
+        return lib.cpBodyGetType(self._body)
+
+    @body_type.setter
+    def body_type(self, body_type: _BodyType) -> None:
+        if body_type != Body.DYNAMIC:
+            for c in self.constraints:
+                assert (c.a != self and c.b.body_type == Body.DYNAMIC) or (
+                    c.b != self and c.a.body_type == Body.DYNAMIC
+                ), "Cannot set a non-dynamic body type when Body is connected to a constraint {c} with a non-dynamic other body."
+
+        lib.cpBodySetType(self._body, body_type)
 
     def each_arbiter(
         self,
-        func: Callable[..., None],  # TODO: Fix me once PEP 612 is ready
+        func: Callable[..., None],  # TODO: Fix me once PEP 612 is ready (Python 3.10)
         *args: Any,
         **kwargs: Any,
     ) -> None:
@@ -655,7 +640,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         lib.cpBodyEachArbiter(self._body, lib.ext_cpBodyArbiterIteratorFunc, data)
 
     @property
-    def constraints(self) -> Set["Constraint"]:
+    def constraints(self) -> KeysView["Constraint"]:
         """Get the constraints this body is attached to.
 
         It is not possible to detach a body from a constraint. The only way is
@@ -666,17 +651,25 @@ class Body(PickleMixin, TypingAttrMixing, object):
         collected it will automatically be removed from this collection as
         well.
         """
-        return set(self._constraints)
+        return WeakKeysView(self._constraints)
 
     @property
-    def shapes(self) -> Set["Shape"]:
+    def shapes(self) -> KeysView["Shape"]:
         """Get the shapes attached to this body.
 
-        The body only keeps a weak reference to the shapes and a live
-        body wont prevent GC of the attached shapes"""
-        return set(self._shapes)
+        In case you only have a single shape attached to the body you can
+        unpack it out easily:
 
-    def local_to_world(self, v: Tuple[float, float]) -> Vec2d:
+        >>> from pymunk import Circle
+        >>> b = Body(1)
+        >>> circle = Circle(b, 2)
+        >>> [shape] = b.shapes
+        >>> shape == circle
+        True
+        """
+        return self._shapes.keys()
+
+    def local_to_world(self, v: tuple[float, float]) -> Vec2d:
         """Convert body local coordinates to world space coordinates
 
         Many things are defined in coordinates local to a body meaning that
@@ -689,7 +682,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         v2 = lib.cpBodyLocalToWorld(self._body, v)
         return Vec2d(v2.x, v2.y)
 
-    def world_to_local(self, v: Tuple[float, float]) -> Vec2d:
+    def world_to_local(self, v: tuple[float, float]) -> Vec2d:
         """Convert world space coordinates to body local coordinates
 
         :param v: Vector in world space coordinates
@@ -698,7 +691,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         v2 = lib.cpBodyWorldToLocal(self._body, v)
         return Vec2d(v2.x, v2.y)
 
-    def velocity_at_world_point(self, point: Tuple[float, float]) -> Vec2d:
+    def velocity_at_world_point(self, point: tuple[float, float]) -> Vec2d:
         """Get the absolute velocity of the rigid body at the given world
         point
 
@@ -710,7 +703,7 @@ class Body(PickleMixin, TypingAttrMixing, object):
         v = lib.cpBodyGetVelocityAtWorldPoint(self._body, point)
         return Vec2d(v.x, v.y)
 
-    def velocity_at_local_point(self, point: Tuple[float, float]) -> Vec2d:
+    def velocity_at_local_point(self, point: tuple[float, float]) -> Vec2d:
         """Get the absolute velocity of the rigid body at the given body
         local point
         """
