@@ -28,7 +28,7 @@ body, or they can be more abstract like the gear joint or motors.
 
 This submodule contain all the constraints that are supported by Pymunk.
 
-All the constraints support copy and pickle from the standard library. Custom 
+All the constraints support copy and pickle from the standard library. Custom
 properties set on a constraint will also be copied/pickled.
 
 Chipmunk has a good overview of the different constraint on youtube which
@@ -68,15 +68,15 @@ __all__ = [
     "SimpleMotor",
 ]
 
-from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Any, Callable, Optional, Union
 
 if TYPE_CHECKING:
-    from .body import Body
     from .space import Space
 
 from ._chipmunk_cffi import ffi, lib
 from ._pickle import PickleMixin
 from ._typing_attr import TypingAttrMixing
+from .body import Body
 from .vec2d import Vec2d
 
 _TorqueFunc = Callable[["DampedRotarySpring", float], float]
@@ -102,6 +102,9 @@ class Constraint(PickleMixin, TypingAttrMixing, object):
     _pre_solve_func: Optional[Callable[["Constraint", "Space"], None]] = None
     _post_solve_func: Optional[Callable[["Constraint", "Space"], None]] = None
 
+    _a: "Body"
+    _b: "Body"
+
     def __init__(self, constraint: ffi.CData) -> None:
         self._constraint = constraint
 
@@ -120,32 +123,22 @@ class Constraint(PickleMixin, TypingAttrMixing, object):
         self._data_handle = d  # to prevent gc to collect the handle
         lib.cpConstraintSetUserData(self._constraint, d)
 
-    def _get_max_force(self) -> float:
-        return lib.cpConstraintGetMaxForce(self._constraint)
-
-    def _set_max_force(self, f: float) -> None:
-        lib.cpConstraintSetMaxForce(self._constraint, f)
-
-    max_force = property(
-        _get_max_force,
-        _set_max_force,
-        doc="""The maximum force that the constraint can use to act on the two
+    @property
+    def max_force(self) -> float:
+        """The maximum force that the constraint can use to act on the two
         bodies.
 
         Defaults to infinity
-        """,
-    )
+        """
+        return lib.cpConstraintGetMaxForce(self._constraint)
 
-    def _get_error_bias(self) -> float:
-        return lib.cpConstraintGetErrorBias(self._constraint)
+    @max_force.setter
+    def max_force(self, f: float) -> None:
+        lib.cpConstraintSetMaxForce(self._constraint, f)
 
-    def _set_error_bias(self, error_bias: float) -> None:
-        lib.cpConstraintSetErrorBias(self._constraint, error_bias)
-
-    error_bias = property(
-        _get_error_bias,
-        _set_error_bias,
-        doc="""The percentage of joint error that remains unfixed after a
+    @property
+    def error_bias(self) -> float:
+        """The percentage of joint error that remains unfixed after a
         second.
 
         This works exactly the same as the collision bias property of a space,
@@ -154,42 +147,40 @@ class Constraint(PickleMixin, TypingAttrMixing, object):
 
         Defaults to pow(1.0 - 0.1, 60.0) meaning that it will correct 10% of
         the error every 1/60th of a second.
-        """,
-    )
+        """
+        return lib.cpConstraintGetErrorBias(self._constraint)
 
-    def _get_max_bias(self) -> float:
-        return lib.cpConstraintGetMaxBias(self._constraint)
+    @error_bias.setter
+    def error_bias(self, error_bias: float) -> None:
+        lib.cpConstraintSetErrorBias(self._constraint, error_bias)
 
-    def _set_max_bias(self, max_bias: float) -> None:
-        lib.cpConstraintSetMaxBias(self._constraint, max_bias)
-
-    max_bias = property(
-        _get_max_bias,
-        _set_max_bias,
-        doc="""The maximum speed at which the constraint can apply error
+    @property
+    def max_bias(self) -> float:
+        """The maximum speed at which the constraint can apply error
         correction.
 
         Defaults to infinity
-        """,
-    )
+        """
+        return lib.cpConstraintGetMaxBias(self._constraint)
 
-    def _get_collide_bodies(self) -> bool:
-        return lib.cpConstraintGetCollideBodies(self._constraint)
+    @max_bias.setter
+    def max_bias(self, max_bias: float) -> None:
+        lib.cpConstraintSetMaxBias(self._constraint, max_bias)
 
-    def _set_collide_bodies(self, collide_bodies: bool) -> None:
-        lib.cpConstraintSetCollideBodies(self._constraint, collide_bodies)
-
-    collide_bodies = property(
-        _get_collide_bodies,
-        _set_collide_bodies,
-        doc="""Constraints can be used for filtering collisions too.
+    @property
+    def collide_bodies(self) -> bool:
+        """Constraints can be used for filtering collisions too.
 
         When two bodies collide, Pymunk ignores the collisions if this property
         is set to False on any constraint that connects the two bodies.
         Defaults to True. This can be used to create a chain that self
         collides, but adjacent links in the chain do not collide.
-        """,
-    )
+        """
+        return lib.cpConstraintGetCollideBodies(self._constraint)
+
+    @collide_bodies.setter
+    def collide_bodies(self, collide_bodies: bool) -> None:
+        lib.cpConstraintSetCollideBodies(self._constraint, collide_bodies)
 
     @property
     def impulse(self) -> float:
@@ -213,8 +204,8 @@ class Constraint(PickleMixin, TypingAttrMixing, object):
 
     def activate_bodies(self) -> None:
         """Activate the bodies this constraint is attached to"""
-        self._a.activate()
-        self._b.activate()
+        self.a.activate()
+        self.b.activate()
 
     @property
     def pre_solve(self) -> Optional[Callable[["Constraint", "Space"], None]]:
@@ -275,12 +266,15 @@ class Constraint(PickleMixin, TypingAttrMixing, object):
 
     def _set_bodies(self, a: "Body", b: "Body") -> None:
         assert a is not b
+        assert (
+            a.body_type == Body.DYNAMIC or b.body_type == Body.DYNAMIC
+        ), "At least one of the two bodies attached to a constraint must be DYNAMIC."
         self._a = a
         self._b = b
-        a._constraints.add(self)
-        b._constraints.add(self)
+        a._constraints[self] = None
+        b._constraints[self] = None
 
-    def __getstate__(self) -> Dict[str, List[Tuple[str, Any]]]:
+    def __getstate__(self) -> dict[str, list[tuple[str, Any]]]:
         """Return the state of this object
 
         This method allows the usage of the :mod:`copy` and :mod:`pickle`
@@ -293,7 +287,7 @@ class Constraint(PickleMixin, TypingAttrMixing, object):
 
         return d
 
-    def __setstate__(self, state: Dict[str, List[Tuple[str, Any]]]) -> None:
+    def __setstate__(self, state: dict[str, list[tuple[str, Any]]]) -> None:
         """Unpack this object from a saved state.
 
         This method allows the usage of the :mod:`copy` and :mod:`pickle`
@@ -320,8 +314,8 @@ class PinJoint(Constraint):
         self,
         a: "Body",
         b: "Body",
-        anchor_a: Tuple[float, float] = (0, 0),
-        anchor_b: Tuple[float, float] = (0, 0),
+        anchor_a: tuple[float, float] = (0, 0),
+        anchor_b: tuple[float, float] = (0, 0),
     ) -> None:
         """a and b are the two bodies to connect, and anchor_a and anchor_b are
         the anchor points on those bodies.
@@ -339,7 +333,7 @@ class PinJoint(Constraint):
         v = lib.cpPinJointGetAnchorA(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_a(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_a(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpPinJointSetAnchorA(self._constraint, anchor)
 
@@ -349,19 +343,19 @@ class PinJoint(Constraint):
         v = lib.cpPinJointGetAnchorB(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_b(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_b(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpPinJointSetAnchorB(self._constraint, anchor)
 
     anchor_b = property(_get_anchor_b, _set_anchor_b)
 
-    def _get_distance(self) -> float:
+    @property
+    def distance(self) -> float:
         return lib.cpPinJointGetDist(self._constraint)
 
-    def _set_distance(self, distance: float) -> None:
+    @distance.setter
+    def distance(self, distance: float) -> None:
         lib.cpPinJointSetDist(self._constraint, distance)
-
-    distance = property(_get_distance, _set_distance)
 
 
 class SlideJoint(Constraint):
@@ -382,8 +376,8 @@ class SlideJoint(Constraint):
         self,
         a: "Body",
         b: "Body",
-        anchor_a: Tuple[float, float],
-        anchor_b: Tuple[float, float],
+        anchor_a: tuple[float, float],
+        anchor_b: tuple[float, float],
         min: float,
         max: float,
     ) -> None:
@@ -402,7 +396,7 @@ class SlideJoint(Constraint):
         v = lib.cpSlideJointGetAnchorA(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_a(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_a(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpSlideJointSetAnchorA(self._constraint, anchor)
 
@@ -412,27 +406,27 @@ class SlideJoint(Constraint):
         v = lib.cpSlideJointGetAnchorB(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_b(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_b(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpSlideJointSetAnchorB(self._constraint, anchor)
 
     anchor_b = property(_get_anchor_b, _set_anchor_b)
 
-    def _get_min(self) -> float:
+    @property
+    def min(self) -> float:
         return lib.cpSlideJointGetMin(self._constraint)
 
-    def _set_min(self, min: float) -> None:
+    @min.setter
+    def min(self, min: float) -> None:
         lib.cpSlideJointSetMin(self._constraint, min)
 
-    min = property(_get_min, _set_min)
-
-    def _get_max(self) -> float:
+    @property
+    def max(self) -> float:
         return lib.cpSlideJointGetMax(self._constraint)
 
-    def _set_max(self, max: float) -> None:
+    @max.setter
+    def max(self, max: float) -> None:
         lib.cpSlideJointSetMax(self._constraint, max)
-
-    max = property(_get_max, _set_max)
 
 
 class PivotJoint(Constraint):
@@ -448,7 +442,7 @@ class PivotJoint(Constraint):
         a: "Body",
         b: "Body",
         *args: Union[
-            Tuple[float, float], Tuple[Tuple[float, float], Tuple[float, float]]
+            tuple[float, float], tuple[tuple[float, float], tuple[float, float]]
         ],
     ) -> None:
         """a and b are the two bodies to connect, and pivot is the point in
@@ -486,7 +480,7 @@ class PivotJoint(Constraint):
         v = lib.cpPivotJointGetAnchorA(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_a(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_a(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpPivotJointSetAnchorA(self._constraint, anchor)
 
@@ -496,7 +490,7 @@ class PivotJoint(Constraint):
         v = lib.cpPivotJointGetAnchorB(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_b(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_b(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpPivotJointSetAnchorB(self._constraint, anchor)
 
@@ -519,9 +513,9 @@ class GrooveJoint(Constraint):
         self,
         a: "Body",
         b: "Body",
-        groove_a: Tuple[float, float],
-        groove_b: Tuple[float, float],
-        anchor_b: Tuple[float, float],
+        groove_a: tuple[float, float],
+        groove_b: tuple[float, float],
+        anchor_b: tuple[float, float],
     ) -> None:
         """The groove goes from groove_a to groove_b on body a, and the pivot
         is attached to anchor_b on body b.
@@ -540,7 +534,7 @@ class GrooveJoint(Constraint):
         v = lib.cpGrooveJointGetAnchorB(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_b(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_b(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpGrooveJointSetAnchorB(self._constraint, anchor)
 
@@ -550,7 +544,7 @@ class GrooveJoint(Constraint):
         v = lib.cpGrooveJointGetGrooveA(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_groove_a(self, groove: Tuple[float, float]) -> None:
+    def _set_groove_a(self, groove: tuple[float, float]) -> None:
         assert len(groove) == 2
         lib.cpGrooveJointSetGrooveA(self._constraint, groove)
 
@@ -560,7 +554,7 @@ class GrooveJoint(Constraint):
         v = lib.cpGrooveJointGetGrooveB(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_groove_b(self, groove: Tuple[float, float]) -> None:
+    def _set_groove_b(self, groove: tuple[float, float]) -> None:
         assert len(groove) == 2
         lib.cpGrooveJointSetGrooveB(self._constraint, groove)
 
@@ -587,8 +581,8 @@ class DampedSpring(Constraint):
         self,
         a: "Body",
         b: "Body",
-        anchor_a: Tuple[float, float],
-        anchor_b: Tuple[float, float],
+        anchor_a: tuple[float, float],
+        anchor_b: tuple[float, float],
         rest_length: float,
         stiffness: float,
         damping: float,
@@ -623,7 +617,7 @@ class DampedSpring(Constraint):
         v = lib.cpDampedSpringGetAnchorA(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_a(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_a(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpDampedSpringSetAnchorA(self._constraint, anchor)
 
@@ -633,45 +627,38 @@ class DampedSpring(Constraint):
         v = lib.cpDampedSpringGetAnchorB(self._constraint)
         return Vec2d(v.x, v.y)
 
-    def _set_anchor_b(self, anchor: Tuple[float, float]) -> None:
+    def _set_anchor_b(self, anchor: tuple[float, float]) -> None:
         assert len(anchor) == 2
         lib.cpDampedSpringSetAnchorB(self._constraint, anchor)
 
     anchor_b = property(_get_anchor_b, _set_anchor_b)
 
-    def _get_rest_length(self) -> float:
+    @property
+    def rest_length(self) -> float:
+        """The distance the spring wants to be."""
         return lib.cpDampedSpringGetRestLength(self._constraint)
 
-    def _set_rest_length(self, rest_length: float) -> None:
+    @rest_length.setter
+    def rest_length(self, rest_length: float) -> None:
         lib.cpDampedSpringSetRestLength(self._constraint, rest_length)
 
-    rest_length = property(
-        _get_rest_length,
-        _set_rest_length,
-        doc="""The distance the spring wants to be.""",
-    )
-
-    def _get_stiffness(self) -> float:
+    @property
+    def stiffness(self) -> float:
+        """The spring constant (Young's modulus)."""
         return lib.cpDampedSpringGetStiffness(self._constraint)
 
-    def _set_stiffness(self, stiffness: float) -> None:
+    @stiffness.setter
+    def stiffness(self, stiffness: float) -> None:
         lib.cpDampedSpringSetStiffness(self._constraint, stiffness)
 
-    stiffness = property(
-        _get_stiffness, _set_stiffness, doc="""The spring constant (Young's modulus)."""
-    )
-
-    def _get_damping(self) -> float:
+    @property
+    def damping(self) -> float:
+        """How soft to make the damping of the spring."""
         return lib.cpDampedSpringGetDamping(self._constraint)
 
-    def _set_damping(self, damping: float) -> None:
+    @damping.setter
+    def damping(self, damping: float) -> None:
         lib.cpDampedSpringSetDamping(self._constraint, damping)
-
-    damping = property(
-        _get_damping,
-        _set_damping,
-        doc="""How soft to make the damping of the spring.""",
-    )
 
     @staticmethod
     def spring_force(spring: "DampedSpring", dist: float) -> float:
@@ -734,39 +721,32 @@ class DampedRotarySpring(Constraint):
         )
         self._init(a, b, _constraint)
 
-    def _get_rest_angle(self) -> float:
+    @property
+    def rest_angle(self) -> float:
+        """The relative angle in radians that the bodies want to have"""
         return lib.cpDampedRotarySpringGetRestAngle(self._constraint)
 
-    def _set_rest_angle(self, rest_angle: float) -> None:
+    @rest_angle.setter
+    def rest_angle(self, rest_angle: float) -> None:
         lib.cpDampedRotarySpringSetRestAngle(self._constraint, rest_angle)
 
-    rest_angle = property(
-        _get_rest_angle,
-        _set_rest_angle,
-        doc="""The relative angle in radians that the bodies want to have""",
-    )
-
-    def _get_stiffness(self) -> float:
+    @property
+    def stiffness(self) -> float:
+        """The spring constant (Young's modulus)."""
         return lib.cpDampedRotarySpringGetStiffness(self._constraint)
 
-    def _set_stiffness(self, stiffness: float) -> None:
+    @stiffness.setter
+    def stiffness(self, stiffness: float) -> None:
         lib.cpDampedRotarySpringSetStiffness(self._constraint, stiffness)
 
-    stiffness = property(
-        _get_stiffness, _set_stiffness, doc="""The spring constant (Young's modulus)."""
-    )
-
-    def _get_damping(self) -> float:
+    @property
+    def damping(self) -> float:
+        """How soft to make the damping of the spring."""
         return lib.cpDampedRotarySpringGetDamping(self._constraint)
 
-    def _set_damping(self, damping: float) -> None:
+    @damping.setter
+    def damping(self, damping: float) -> None:
         lib.cpDampedRotarySpringSetDamping(self._constraint, damping)
-
-    damping = property(
-        _get_damping,
-        _set_damping,
-        doc="""How soft to make the damping of the spring.""",
-    )
 
     @staticmethod
     def spring_torque(spring: "DampedRotarySpring", relative_angle: float) -> float:
@@ -816,21 +796,21 @@ class RotaryLimitJoint(Constraint):
         _constraint = lib.cpRotaryLimitJointNew(a._body, b._body, min, max)
         self._init(a, b, _constraint)
 
-    def _get_min(self) -> float:
+    @property
+    def min(self) -> float:
         return lib.cpRotaryLimitJointGetMin(self._constraint)
 
-    def _set_min(self, min: float) -> None:
+    @min.setter
+    def min(self, min: float) -> None:
         lib.cpRotaryLimitJointSetMin(self._constraint, min)
 
-    min = property(_get_min, _set_min)
-
-    def _get_max(self) -> float:
+    @property
+    def max(self) -> float:
         return lib.cpRotaryLimitJointGetMax(self._constraint)
 
-    def _set_max(self, max: float) -> None:
+    @max.setter
+    def max(self, max: float) -> None:
         lib.cpRotaryLimitJointSetMax(self._constraint, max)
-
-    max = property(_get_max, _set_max)
 
 
 class RatchetJoint(Constraint):
@@ -847,29 +827,29 @@ class RatchetJoint(Constraint):
         _constraint = lib.cpRatchetJointNew(a._body, b._body, phase, ratchet)
         self._init(a, b, _constraint)
 
-    def _get_angle(self) -> float:
+    @property
+    def angle(self) -> float:
         return lib.cpRatchetJointGetAngle(self._constraint)
 
-    def _set_angle(self, angle: float) -> None:
+    @angle.setter
+    def angle(self, angle: float) -> None:
         lib.cpRatchetJointSetAngle(self._constraint, angle)
 
-    angle = property(_get_angle, _set_angle)
-
-    def _get_phase(self) -> float:
+    @property
+    def phase(self) -> float:
         return lib.cpRatchetJointGetPhase(self._constraint)
 
-    def _set_phase(self, phase: float) -> None:
+    @phase.setter
+    def phase(self, phase: float) -> None:
         lib.cpRatchetJointSetPhase(self._constraint, phase)
 
-    phase = property(_get_phase, _set_phase)
-
-    def _get_ratchet(self) -> float:
+    @property
+    def ratchet(self) -> float:
         return lib.cpRatchetJointGetRatchet(self._constraint)
 
-    def _set_ratchet(self, ratchet: float) -> None:
+    @ratchet.setter
+    def ratchet(self, ratchet: float) -> None:
         lib.cpRatchetJointSetRatchet(self._constraint, ratchet)
-
-    ratchet = property(_get_ratchet, _set_ratchet)
 
 
 class GearJoint(Constraint):
@@ -887,21 +867,21 @@ class GearJoint(Constraint):
         _constraint = lib.cpGearJointNew(a._body, b._body, phase, ratio)
         self._init(a, b, _constraint)
 
-    def _get_phase(self) -> float:
+    @property
+    def phase(self) -> float:
         return lib.cpGearJointGetPhase(self._constraint)
 
-    def _set_phase(self, phase: float) -> None:
+    @phase.setter
+    def phase(self, phase: float) -> None:
         lib.cpGearJointSetPhase(self._constraint, phase)
 
-    phase = property(_get_phase, _set_phase)
-
-    def _get_ratio(self) -> float:
+    @property
+    def ratio(self) -> float:
         return lib.cpGearJointGetRatio(self._constraint)
 
-    def _set_ratio(self, ratio: float) -> None:
+    @ratio.setter
+    def ratio(self, ratio: float) -> None:
         lib.cpGearJointSetRatio(self._constraint, ratio)
-
-    ratio = property(_get_ratio, _set_ratio)
 
 
 class SimpleMotor(Constraint):
@@ -919,12 +899,11 @@ class SimpleMotor(Constraint):
         _constraint = lib.cpSimpleMotorNew(a._body, b._body, rate)
         self._init(a, b, _constraint)
 
-    def _get_rate(self) -> float:
+    @property
+    def rate(self) -> float:
+        """The desired relative angular velocity"""
         return lib.cpSimpleMotorGetRate(self._constraint)
 
-    def _set_rate(self, rate: float) -> None:
+    @rate.setter
+    def rate(self, rate: float) -> None:
         lib.cpSimpleMotorSetRate(self._constraint, rate)
-
-    rate = property(
-        _get_rate, _set_rate, doc="""The desired relative angular velocity"""
-    )
