@@ -1,6 +1,7 @@
 __docformat__ = "reStructuredText"
 
-from typing import TYPE_CHECKING, List, Optional, Sequence, Tuple
+import weakref
+from typing import TYPE_CHECKING, Optional, Sequence
 
 if TYPE_CHECKING:
     from .body import Body
@@ -10,6 +11,7 @@ from ._chipmunk_cffi import ffi
 from ._chipmunk_cffi import lib as cp
 from ._pickle import PickleMixin, _State
 from ._typing_attr import TypingAttrMixing
+from ._util import _dead_ref
 from .bb import BB
 from .contact_point_set import ContactPointSet
 from .query_info import PointQueryInfo, SegmentQueryInfo
@@ -41,80 +43,62 @@ class Shape(PickleMixin, TypingAttrMixing, object):
     ]
     _pickle_attrs_skip = PickleMixin._pickle_attrs_skip + ["mass", "density"]
 
-    _space = None  # Weak ref to the space holding this body (if any)
-
-    _id_counter = 1
+    _space: weakref.ref["Space"] = _dead_ref
 
     def __init__(self, shape: "Shape") -> None:
         self._shape = shape
-        self._body: Optional["Body"] = shape.body
+        assert shape.body != None
+        self._body: weakref.ref["Body"] = weakref.ref(shape.body)
 
     def _init(self, body: Optional["Body"], _shape: ffi.CData) -> None:
-        self._body = body
 
         if body is not None:
-            body._shapes.add(self)
+            self._body = weakref.ref(body)
+            body._shapes[self] = None
+        else:
+            self._body = _dead_ref
 
         def shapefree(cp_shape: ffi.CData) -> None:
             cp_space = cp.cpShapeGetSpace(cp_shape)
             if cp_space != ffi.NULL:
                 cp.cpSpaceRemoveShape(cp_space, cp_shape)
 
-            cp_body = cp.cpShapeGetBody(cp_shape)
-            if cp_body != ffi.NULL:
-                cp.cpShapeSetBody(cp_shape, ffi.NULL)
+            # cp_body = cp.cpShapeGetBody(cp_shape)
+            # if cp_body != ffi.NULL:
+            #     cp.cpShapeSetBody(cp_shape, ffi.NULL)
             cp.cpShapeFree(cp_shape)
 
         self._shape = ffi.gc(_shape, shapefree)
-        self._set_id()
+        self._h = ffi.new_handle(self)  # to prevent GC of the handle
+        cp.cpShapeSetUserData(self._shape, self._h)
 
     @property
-    def _id(self) -> int:
-        """Unique id of the Shape.
+    def mass(self) -> float:
+        """The mass of this shape.
 
-        .. note::
-            Experimental API. Likely to change in future major, minor orpoint
-            releases.
+        This is useful when you let Pymunk calculate the total mass and inertia
+        of a body from the shapes attached to it. (Instead of setting the body
+        mass and inertia directly)
         """
-        return int(ffi.cast("int", cp.cpShapeGetUserData(self._shape)))
-
-    def _set_id(self) -> None:
-        cp.cpShapeSetUserData(self._shape, ffi.cast("cpDataPointer", Shape._id_counter))
-        Shape._id_counter += 1
-
-    def _get_mass(self) -> float:
         return cp.cpShapeGetMass(self._shape)
 
-    def _set_mass(self, mass: float) -> None:
+    @mass.setter
+    def mass(self, mass: float) -> None:
         cp.cpShapeSetMass(self._shape, mass)
 
-    mass = property(
-        _get_mass,
-        _set_mass,
-        doc="""The mass of this shape.
+    @property
+    def density(self) -> float:
+        """The density of this shape.
 
         This is useful when you let Pymunk calculate the total mass and inertia
         of a body from the shapes attached to it. (Instead of setting the body
         mass and inertia directly)
-        """,
-    )
-
-    def _get_density(self) -> float:
+        """
         return cp.cpShapeGetDensity(self._shape)
 
-    def _set_density(self, density: float) -> None:
+    @density.setter
+    def density(self, density: float) -> None:
         cp.cpShapeSetDensity(self._shape, density)
-
-    density = property(
-        _get_density,
-        _set_density,
-        doc="""The density of this shape.
-
-        This is useful when you let Pymunk calculate the total mass and inertia
-        of a body from the shapes attached to it. (Instead of setting the body
-        mass and inertia directly)
-        """,
-    )
 
     @property
     def moment(self) -> float:
@@ -132,79 +116,61 @@ class Shape(PickleMixin, TypingAttrMixing, object):
         v = cp.cpShapeGetCenterOfGravity(self._shape)
         return Vec2d(v.x, v.y)
 
-    def _get_sensor(self) -> bool:
-        return bool(cp.cpShapeGetSensor(self._shape))
-
-    def _set_sensor(self, is_sensor: bool) -> None:
-        cp.cpShapeSetSensor(self._shape, is_sensor)
-
-    sensor = property(
-        _get_sensor,
-        _set_sensor,
-        doc="""A boolean value if this shape is a sensor or not.
+    @property
+    def sensor(self) -> bool:
+        """A boolean value if this shape is a sensor or not.
 
         Sensors only call collision callbacks, and never generate real
         collisions.
-        """,
-    )
+        """
+        return bool(cp.cpShapeGetSensor(self._shape))
 
-    def _get_collision_type(self) -> int:
+    @sensor.setter
+    def sensor(self, is_sensor: bool) -> None:
+        cp.cpShapeSetSensor(self._shape, is_sensor)
+
+    @property
+    def collision_type(self) -> int:
+        """User defined collision type for the shape.
+
+        Defaults to 0.
+
+        See the :py:meth:`Space.add_collision_handler` function for more
+        information on when to use this property.
+        """
         return cp.cpShapeGetCollisionType(self._shape)
 
-    def _set_collision_type(self, t: int) -> None:
+    @collision_type.setter
+    def collision_type(self, t: int) -> None:
         cp.cpShapeSetCollisionType(self._shape, t)
 
-    collision_type = property(
-        _get_collision_type,
-        _set_collision_type,
-        doc="""User defined collision type for the shape.
-
-        See :py:meth:`Space.add_collision_handler` function for more
-        information on when to use this property.
-        """,
-    )
-
-    def _get_filter(self) -> ShapeFilter:
+    @property
+    def filter(self) -> ShapeFilter:
+        """Set the collision :py:class:`ShapeFilter` for this shape."""
         f = cp.cpShapeGetFilter(self._shape)
         return ShapeFilter(f.group, f.categories, f.mask)
 
-    def _set_filter(self, f: ShapeFilter) -> None:
+    @filter.setter
+    def filter(self, f: ShapeFilter) -> None:
         cp.cpShapeSetFilter(self._shape, f)
 
-    filter = property(
-        _get_filter,
-        _set_filter,
-        doc="""Set the collision :py:class:`ShapeFilter` for this shape.
-        """,
-    )
-
-    def _get_elasticity(self) -> float:
-        return cp.cpShapeGetElasticity(self._shape)
-
-    def _set_elasticity(self, e: float) -> None:
-        cp.cpShapeSetElasticity(self._shape, e)
-
-    elasticity = property(
-        _get_elasticity,
-        _set_elasticity,
-        doc="""Elasticity of the shape.
+    @property
+    def elasticity(self) -> float:
+        """Elasticity of the shape.
 
         A value of 0.0 gives no bounce, while a value of 1.0 will give a
         'perfect' bounce. However due to inaccuracies in the simulation
         using 1.0 or greater is not recommended.
-        """,
-    )
+        """
+        return cp.cpShapeGetElasticity(self._shape)
 
-    def _get_friction(self) -> float:
-        return cp.cpShapeGetFriction(self._shape)
+    @elasticity.setter
+    def elasticity(self, e: float) -> None:
+        cp.cpShapeSetElasticity(self._shape, e)
 
-    def _set_friction(self, u: float) -> None:
-        cp.cpShapeSetFriction(self._shape, u)
-
-    friction = property(
-        _get_friction,
-        _set_friction,
-        doc="""Friction coefficient.
+    @property
+    def friction(self) -> float:
+        """Friction coefficient.
 
         Pymunk uses the Coulomb friction model, a value of 0.0 is
         frictionless.
@@ -234,14 +200,18 @@ class Shape(PickleMixin, TypingAttrMixing, object):
         Teflon (PTFE)   Teflon  0.04
         Wood            Wood    0.4
         ==============  ======  ========
-        """,
-    )
+        """
+        return cp.cpShapeGetFriction(self._shape)
+
+    @friction.setter
+    def friction(self, u: float) -> None:
+        cp.cpShapeSetFriction(self._shape, u)
 
     def _get_surface_velocity(self) -> Vec2d:
         v = cp.cpShapeGetSurfaceVelocity(self._shape)
         return Vec2d(v.x, v.y)
 
-    def _set_surface_velocity(self, surface_v: Vec2d) -> None:
+    def _set_surface_velocity(self, surface_v: tuple[float, float]) -> None:
         assert len(surface_v) == 2
         cp.cpShapeSetSurfaceVelocity(self._shape, surface_v)
 
@@ -256,24 +226,27 @@ class Shape(PickleMixin, TypingAttrMixing, object):
         """,
     )
 
-    def _get_body(self) -> Optional["Body"]:
-        return self._body
+    @property
+    def body(self) -> Optional["Body"]:
+        """The body this shape is attached to.
 
-    def _set_body(self, body: Optional["Body"]) -> None:
-        if self._body is not None:
-            self._body._shapes.remove(self)
-        body_body = ffi.NULL if body is None else body._body
-        cp.cpShapeSetBody(self._shape, body_body)
+        Can be set to None to indicate that this shape doesnt belong to a body.
+        The shape only holds a weakref to the Body, meaning it wont prevent it
+        from being GCed.
+        """
+        return self._body()
+
+    @body.setter
+    def body(self, body: Optional["Body"]) -> None:
+        if self.body is not None:
+            del self.body._shapes[self]
+        cp_body = ffi.NULL if body is None else body._body
+        cp.cpShapeSetBody(self._shape, cp_body)
         if body is not None:
-            body._shapes.add(self)
-        self._body = body
-
-    body = property(
-        _get_body,
-        _set_body,
-        doc="""The body this shape is attached to. Can be set to None to
-        indicate that this shape doesnt belong to a body.""",
-    )
+            body._shapes[self] = None
+            self._body = weakref.ref(body)
+        else:
+            self._body = _dead_ref
 
     def update(self, transform: Transform) -> BB:
         """Update, cache and return the bounding box of a shape with an
@@ -303,20 +276,20 @@ class Shape(PickleMixin, TypingAttrMixing, object):
         _bb = cp.cpShapeGetBB(self._shape)
         return BB(_bb.l, _bb.b, _bb.r, _bb.t)
 
-    def point_query(self, p: Tuple[float, float]) -> PointQueryInfo:
+    def point_query(self, p: tuple[float, float]) -> PointQueryInfo:
         """Check if the given point lies within the shape.
 
         A negative distance means the point is within the shape.
 
-        :return: Tuple of (distance, info)
+        :return: tuple of (distance, info)
         :rtype: (float, :py:class:`PointQueryInfo`)
         """
         assert len(p) == 2
         info = ffi.new("cpPointQueryInfo *")
         _ = cp.cpShapePointQuery(self._shape, p, info)
 
-        ud = int(ffi.cast("int", cp.cpShapeGetUserData(info.shape)))
-        assert ud == self._id
+        shape = ffi.from_handle(cp.cpShapeGetUserData(info.shape))
+        assert shape == self, "This is a bug in Pymunk. Please report it."
         return PointQueryInfo(
             self,
             Vec2d(info.point.x, info.point.y),
@@ -325,9 +298,11 @@ class Shape(PickleMixin, TypingAttrMixing, object):
         )
 
     def segment_query(
-        self, start: Tuple[float, float], end: Tuple[float, float], radius: float = 0
-    ) -> SegmentQueryInfo:
+        self, start: tuple[float, float], end: tuple[float, float], radius: float = 0
+    ) -> Optional[SegmentQueryInfo]:
         """Check if the line segment from start to end intersects the shape.
+
+        Returns None if it does not intersect
 
         :rtype: :py:class:`SegmentQueryInfo`
         """
@@ -336,8 +311,8 @@ class Shape(PickleMixin, TypingAttrMixing, object):
         info = ffi.new("cpSegmentQueryInfo *")
         r = cp.cpShapeSegmentQuery(self._shape, start, end, radius, info)
         if r:
-            ud = int(ffi.cast("int", cp.cpShapeGetUserData(info.shape)))
-            assert ud == self._id
+            shape = ffi.from_handle(cp.cpShapeGetUserData(info.shape))
+            assert shape == self, "This is a bug in Pymunk. Please report it."
             return SegmentQueryInfo(
                 self,
                 Vec2d(info.point.x, info.point.y),
@@ -345,12 +320,7 @@ class Shape(PickleMixin, TypingAttrMixing, object):
                 info.alpha,
             )
         else:
-            return SegmentQueryInfo(
-                None,
-                Vec2d(info.point.x, info.point.y),
-                Vec2d(info.normal.x, info.normal.y),
-                info.alpha,
-            )
+            return None
 
     def shapes_collide(self, b: "Shape") -> ContactPointSet:
         """Get contact information about this shape and shape b.
@@ -365,13 +335,7 @@ class Shape(PickleMixin, TypingAttrMixing, object):
         """Get the :py:class:`Space` that shape has been added to (or
         None).
         """
-        if self._space is not None:
-            try:
-                return self._space._get_self()  # ugly hack because of weakref
-            except ReferenceError:
-                return None
-        else:
-            return None
+        return self._space()
 
     @property
     def _hashid(self) -> int:
@@ -380,6 +344,13 @@ class Shape(PickleMixin, TypingAttrMixing, object):
     @_hashid.setter
     def _hashid(self, v: int) -> None:
         cp.cpShapeSetHashID(self._shape, v)
+
+    @staticmethod
+    def _from_cp_shape(cp_shape: ffi.CData) -> Optional["Shape"]:
+        """Get Pymunk Shape from a Chipmunk Shape pointer"""
+        if not bool(cp_shape):
+            return None
+        return ffi.from_handle(cp.cpShapeGetUserData(cp_shape))
 
     def __getstate__(self) -> _State:
         """Return the state of this object.
@@ -409,7 +380,7 @@ class Circle(Shape):
         self,
         body: Optional["Body"],
         radius: float,
-        offset: Tuple[float, float] = (0, 0),
+        offset: tuple[float, float] = (0, 0),
     ) -> None:
         """body is the body attach the circle to, offset is the offset from the
         body's center of gravity in body local coordinates.
@@ -439,7 +410,7 @@ class Circle(Shape):
         """The Radius of the circle."""
         return cp.cpCircleShapeGetRadius(self._shape)
 
-    def unsafe_set_offset(self, o: Tuple[float, float]) -> None:
+    def unsafe_set_offset(self, o: tuple[float, float]) -> None:
         """Unsafe set the offset of the circle.
 
         .. note::
@@ -470,8 +441,8 @@ class Segment(Shape):
     def __init__(
         self,
         body: Optional["Body"],
-        a: Tuple[float, float],
-        b: Tuple[float, float],
+        a: tuple[float, float],
+        b: tuple[float, float],
         radius: float,
     ) -> None:
         """Create a Segment.
@@ -492,20 +463,20 @@ class Segment(Shape):
         _shape = cp.cpSegmentShapeNew(body_body, a, b, radius)
         self._init(body, _shape)
 
-    def _get_a(self) -> Vec2d:
+    @property
+    def a(self) -> Vec2d:
+        """The first of the two endpoints for this segment"""
         v = cp.cpSegmentShapeGetA(self._shape)
         return Vec2d(v.x, v.y)
 
-    a = property(_get_a, doc="""The first of the two endpoints for this segment""")
-
-    def _get_b(self) -> Vec2d:
+    @property
+    def b(self) -> Vec2d:
+        """The second of the two endpoints for this segment"""
         v = cp.cpSegmentShapeGetB(self._shape)
         return Vec2d(v.x, v.y)
 
-    b = property(_get_b, doc="""The second of the two endpoints for this segment""")
-
     def unsafe_set_endpoints(
-        self, a: Tuple[float, float], b: Tuple[float, float]
+        self, a: tuple[float, float], b: tuple[float, float]
     ) -> None:
         """Set the two endpoints for this segment.
 
@@ -542,7 +513,7 @@ class Segment(Shape):
         return cp.cpSegmentShapeGetRadius(self._shape)
 
     def set_neighbors(
-        self, prev: Tuple[float, float], next: Tuple[float, float]
+        self, prev: tuple[float, float], next: tuple[float, float]
     ) -> None:
         """When you have a number of segment shapes that are all joined
         together, things can still collide with the "cracks" between the
@@ -563,7 +534,7 @@ class Poly(Shape):
     def __init__(
         self,
         body: Optional["Body"],
-        vertices: Sequence[Tuple[float, float]],
+        vertices: Sequence[tuple[float, float]],
         transform: Optional[Transform] = None,
         radius: float = 0,
     ) -> None:
@@ -571,7 +542,11 @@ class Poly(Shape):
 
         A convex hull will be calculated from the vertexes automatically. Note
         that concave ones will be converted to a convex hull using the Quickhull
-        algorithm.
+        algorithm::
+
+            >>> poly = Poly(None, [(-1,0), (0, -0.5), (1, 0), (0,-1)])
+            >>> poly.get_vertices()
+            [Vec2d(-1.0, 0.0), Vec2d(0.0, -1.0), Vec2d(1.0, 0.0)]
 
         Adding a small radius will bevel the corners and can significantly
         reduce problems where the poly gets stuck on seams in your geometry.
@@ -586,23 +561,21 @@ class Poly(Shape):
 
             Either directly place the vertices like the below example:
 
-            >>> import pymunk
             >>> w, h = 10, 20
             >>> vs = [(-w/2,-h/2), (w/2,-h/2), (w/2,h/2), (-w/2,h/2)]
-            >>> poly_good = pymunk.Poly(None, vs)
+            >>> poly_good = Poly(None, vs)
             >>> print(poly_good.center_of_gravity)
             Vec2d(0.0, 0.0)
 
             Or use a transform to move them:
 
-            >>> import pymunk
             >>> width, height = 10, 20
             >>> vs = [(0, 0), (width, 0), (width, height), (0, height)]
-            >>> poly_bad = pymunk.Poly(None, vs)
+            >>> poly_bad = Poly(None, vs)
             >>> print(poly_bad.center_of_gravity)
             Vec2d(5.0, 10.0)
-            >>> t = pymunk.Transform(tx=-width/2, ty=-height/2)
-            >>> poly_good = pymunk.Poly(None, vs, transform=t)
+            >>> t = Transform(tx=-width/2, ty=-height/2)
+            >>> poly_good = Poly(None, vs, transform=t)
             >>> print(poly_good.center_of_gravity)
             Vec2d(0.0, 0.0)
 
@@ -643,7 +616,7 @@ class Poly(Shape):
 
     @staticmethod
     def create_box(
-        body: Optional["Body"], size: Tuple[float, float] = (10, 10), radius: float = 0
+        body: Optional["Body"], size: tuple[float, float] = (10, 10), radius: float = 0
     ) -> "Poly":
         """Convenience function to create a box with given width and height.
 
@@ -692,7 +665,7 @@ class Poly(Shape):
 
         return self
 
-    def get_vertices(self) -> List[Vec2d]:
+    def get_vertices(self) -> list[Vec2d]:
         """Get the vertices in local coordinates for the polygon.
 
         If you need the vertices in world coordinates then the vertices can be
@@ -723,7 +696,7 @@ class Poly(Shape):
 
     def unsafe_set_vertices(
         self,
-        vertices: Sequence[Tuple[float, float]],
+        vertices: Sequence[tuple[float, float]],
         transform: Optional[Transform] = None,
     ) -> None:
         """Unsafe set the vertices of the poly.
