@@ -1,19 +1,10 @@
 __docformat__ = "reStructuredText"
 
+import math
 import platform
 import weakref
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    Dict,
-    Hashable,
-    List,
-    Optional,
-    Set,
-    Tuple,
-    Union,
-)
+from collections.abc import KeysView
+from typing import TYPE_CHECKING, Any, Callable, Hashable, Optional, Union
 
 from pymunk.constraints import Constraint
 from pymunk.shape_filter import ShapeFilter
@@ -22,13 +13,11 @@ from pymunk.space_debug_draw_options import SpaceDebugDrawOptions
 from . import _version
 from ._callbacks import *
 from ._chipmunk_cffi import ffi, lib
-
-cp = lib
-
+from ._collision_handler import CollisionHandler, _CollisionCallback
 from ._pickle import PickleMixin, _State
-from .arbiter import _arbiter_from_dict, _arbiter_to_dict
+from ._util import _dead_ref
+from .arbiter import Arbiter, _arbiter_from_dict, _arbiter_to_dict
 from .body import Body
-from .collision_handler import CollisionHandler
 from .query_info import PointQueryInfo, SegmentQueryInfo, ShapeQueryInfo
 from .shapes import Shape
 from .vec2d import Vec2d
@@ -90,16 +79,18 @@ class Space(PickleMixin, object):
         self.threaded = threaded and platform.system() != "Windows"
 
         if self.threaded:
-            cp_space = cp.cpHastySpaceNew()
-            freefunc = cp.cpHastySpaceFree
+            cp_space = lib.cpHastySpaceNew()
+            freefunc = lib.cpHastySpaceFree
         else:
-            cp_space = cp.cpSpaceNew()
-            freefunc = cp.cpSpaceFree
+            cp_space = lib.cpSpaceNew()
+            freefunc = lib.cpSpaceFree
 
         def spacefree(cp_space: ffi.CData) -> None:
-            cp_shapes: List[Shape] = []
+            cp_shapes: list[Shape] = []
             cp_shapes_h = ffi.new_handle(cp_shapes)
-            cp.cpSpaceEachShape(cp_space, lib.ext_cpSpaceShapeIteratorFunc, cp_shapes_h)
+            lib.cpSpaceEachShape(
+                cp_space, lib.ext_cpSpaceShapeIteratorFunc, cp_shapes_h
+            )
 
             for cp_shape in cp_shapes:
                 cp_space = lib.cpShapeGetSpace(cp_shape)
@@ -107,18 +98,18 @@ class Space(PickleMixin, object):
                 lib.cpSpaceRemoveShape(cp_space, cp_shape)
                 lib.cpShapeSetBody(cp_shape, ffi.NULL)
 
-            cp_constraints: List[Constraint] = []
+            cp_constraints: list[Constraint] = []
             cp_constraints_h = ffi.new_handle(cp_constraints)
-            cp.cpSpaceEachConstraint(
+            lib.cpSpaceEachConstraint(
                 cp_space, lib.ext_cpSpaceConstraintIteratorFunc, cp_constraints_h
             )
             for cp_constraint in cp_constraints:
                 cp_space = lib.cpConstraintGetSpace(cp_constraint)
                 lib.cpSpaceRemoveConstraint(cp_space, cp_constraint)
 
-            cp_bodys: List[Body] = []
+            cp_bodys: list[Body] = []
             cp_bodys_h = ffi.new_handle(cp_bodys)
-            cp.cpSpaceEachBody(cp_space, lib.ext_cpSpaceBodyIteratorFunc, cp_bodys_h)
+            lib.cpSpaceEachBody(cp_space, lib.ext_cpSpaceBodyIteratorFunc, cp_bodys_h)
             for cp_body in cp_bodys:
                 cp_space = lib.cpBodyGetSpace(cp_body)
                 lib.cpSpaceRemoveBody(cp_space, cp_body)
@@ -127,47 +118,70 @@ class Space(PickleMixin, object):
 
         self._space = ffi.gc(cp_space, spacefree)
 
-        self._handlers: Dict[Any, CollisionHandler] = (
+        self._handlers: dict[Any, CollisionHandler] = (
             {}
         )  # To prevent the gc to collect the callbacks.
 
-        self._post_step_callbacks: Dict[Any, Callable[["Space"], None]] = {}
-        self._removed_shapes: Dict[int, Shape] = {}
+        self._post_step_callbacks: dict[Any, Callable[["Space"], None]] = {}
+        self._removed_shapes: dict[Shape, None] = {}
 
-        self._shapes: Dict[int, Shape] = {}
-        self._bodies: Dict[Body, None] = {}
+        self._shapes: dict[Shape, None] = {}
+        self._bodies: dict[Body, None] = {}
         self._static_body: Optional[Body] = None
-        self._constraints: Dict[Constraint, None] = {}
+        self._constraints: dict[Constraint, None] = {}
 
         self._locked = False
 
-        self._add_later: Set[_AddableObjects] = set()
-        self._remove_later: Set[_AddableObjects] = set()
-
-    def _get_self(self) -> "Space":
-        return self
+        self._add_later: set[_AddableObjects] = set()
+        self._remove_later: set[_AddableObjects] = set()
+        self._bodies_to_check: set[Body] = set()
 
     @property
-    def shapes(self) -> List[Shape]:
-        """A list of all the shapes added to this space
+    def shapes(self) -> KeysView[Shape]:
+        """The shapes added to this space returned as a KeysView.
 
-        (includes both static and non-static)
+        Since its a view that is returned it will update as shapes are
+        added.
+
+        >>> import pymunk
+        >>> s = pymunk.Space()
+        >>> s.add(pymunk.Circle(s.static_body, 1))
+        >>> shapes_view = s.shapes
+        >>> len(shapes_view)
+        1
+        >>> s.add(pymunk.Circle(s.static_body, 2))
+        >>> len(shapes_view)
+        2
         """
-        return list(self._shapes.values())
+        return self._shapes.keys()
 
     @property
-    def bodies(self) -> List[Body]:
-        """A list of the bodies added to this space"""
-        return list(self._bodies)
+    def bodies(self) -> KeysView[Body]:
+        """The bodies added to this space returned as a KeysView.
+
+        This includes both static and non-static bodies added to the Space.
+        Since its a view that is returned it will update as bodies are added:
+
+        >>> import pymunk
+        >>> s = pymunk.Space()
+        >>> s.add(pymunk.Body())
+        >>> bodies_view = s.bodies
+        >>> len(bodies_view)
+        1
+        >>> s.add(pymunk.Body())
+        >>> len(bodies_view)
+        2
+        """
+        return self._bodies.keys()
 
     @property
-    def constraints(self) -> List[Constraint]:
-        """A list of the constraints added to this space"""
-        return list(self._constraints)
+    def constraints(self) -> KeysView[Constraint]:
+        """The constraints added to this space as a KeysView."""
+        return self._constraints.keys()
 
     def _setup_static_body(self, static_body: Body) -> None:
-        static_body._space = weakref.proxy(self)
-        cp.cpSpaceAddBody(self._space, static_body._body)
+        static_body._space = weakref.ref(self)
+        lib.cpSpaceAddBody(self._space, static_body._body)
 
     @property
     def static_body(self) -> Body:
@@ -181,22 +195,15 @@ class Space(PickleMixin, object):
             self._setup_static_body(self._static_body)
             # self.add(self._static_body)
 
-            # b = cp.cpSpaceGetStaticBody(self._space)
+            # b = lib.cpSpaceGetStaticBody(self._space)
             # self._static_body = Body._init_with_body(b)
             # self._static_body._space = self
             # assert self._static_body is not None
         return self._static_body
 
-    def _set_iterations(self, value: int) -> None:
-        cp.cpSpaceSetIterations(self._space, value)
-
-    def _get_iterations(self) -> int:
-        return cp.cpSpaceGetIterations(self._space)
-
-    iterations = property(
-        _get_iterations,
-        _set_iterations,
-        doc="""Iterations allow you to control the accuracy of the solver.
+    @property
+    def iterations(self) -> int:
+        """Iterations allow you to control the accuracy of the solver.
 
         Defaults to 10.
 
@@ -212,15 +219,19 @@ class Space(PickleMixin, object):
         the number of iterations lets you balance between CPU usage and the
         accuracy of the physics. Pymunk's default of 10 iterations is
         sufficient for most simple games.
-        """,
-    )
+        """
+        return lib.cpSpaceGetIterations(self._space)
 
-    def _set_gravity(self, gravity_vector: Tuple[float, float]) -> None:
+    @iterations.setter
+    def iterations(self, value: int) -> None:
+        lib.cpSpaceSetIterations(self._space, value)
+
+    def _set_gravity(self, gravity_vector: tuple[float, float]) -> None:
         assert len(gravity_vector) == 2
-        cp.cpSpaceSetGravity(self._space, gravity_vector)
+        lib.cpSpaceSetGravity(self._space, gravity_vector)
 
     def _get_gravity(self) -> Vec2d:
-        v = cp.cpSpaceGetGravity(self._space)
+        v = lib.cpSpaceGetGravity(self._space)
         return Vec2d(v.x, v.y)
 
     gravity = property(
@@ -234,81 +245,62 @@ class Space(PickleMixin, object):
         """,
     )
 
-    def _set_damping(self, damping: float) -> None:
-        cp.cpSpaceSetDamping(self._space, damping)
-
-    def _get_damping(self) -> float:
-        return cp.cpSpaceGetDamping(self._space)
-
-    damping = property(
-        _get_damping,
-        _set_damping,
-        doc="""Amount of simple damping to apply to the space.
+    @property
+    def damping(self) -> float:
+        """Amount of simple damping to apply to the space.
 
         A value of 0.9 means that each body will lose 10% of its velocity per
         second. Defaults to 1. Like gravity, it can be overridden on a per
         body basis.
-        """,
-    )
+        """
+        return lib.cpSpaceGetDamping(self._space)
 
-    def _set_idle_speed_threshold(self, idle_speed_threshold: float) -> None:
-        cp.cpSpaceSetIdleSpeedThreshold(self._space, idle_speed_threshold)
+    @damping.setter
+    def damping(self, damping: float) -> None:
+        lib.cpSpaceSetDamping(self._space, damping)
 
-    def _get_idle_speed_threshold(self) -> float:
-        return cp.cpSpaceGetIdleSpeedThreshold(self._space)
-
-    idle_speed_threshold = property(
-        _get_idle_speed_threshold,
-        _set_idle_speed_threshold,
-        doc="""Speed threshold for a body to be considered idle.
+    @property
+    def idle_speed_threshold(self) -> float:
+        """Speed threshold for a body to be considered idle.
 
         The default value of 0 means the space estimates a good threshold
         based on gravity.
-        """,
-    )
+        """
+        return lib.cpSpaceGetIdleSpeedThreshold(self._space)
 
-    def _set_sleep_time_threshold(self, sleep_time_threshold: float) -> None:
-        cp.cpSpaceSetSleepTimeThreshold(self._space, sleep_time_threshold)
+    @idle_speed_threshold.setter
+    def idle_speed_threshold(self, idle_speed_threshold: float) -> None:
+        lib.cpSpaceSetIdleSpeedThreshold(self._space, idle_speed_threshold)
 
-    def _get_sleep_time_threshold(self) -> float:
-        return cp.cpSpaceGetSleepTimeThreshold(self._space)
-
-    sleep_time_threshold = property(
-        _get_sleep_time_threshold,
-        _set_sleep_time_threshold,
-        doc="""Time a group of bodies must remain idle in order to fall
+    @property
+    def sleep_time_threshold(self) -> float:
+        """Time a group of bodies must remain idle in order to fall
         asleep.
 
         The default value of `inf` disables the sleeping algorithm.
-        """,
-    )
+        """
+        return lib.cpSpaceGetSleepTimeThreshold(self._space)
 
-    def _set_collision_slop(self, collision_slop: float) -> None:
-        cp.cpSpaceSetCollisionSlop(self._space, collision_slop)
+    @sleep_time_threshold.setter
+    def sleep_time_threshold(self, sleep_time_threshold: float) -> None:
+        lib.cpSpaceSetSleepTimeThreshold(self._space, sleep_time_threshold)
 
-    def _get_collision_slop(self) -> float:
-        return cp.cpSpaceGetCollisionSlop(self._space)
-
-    collision_slop = property(
-        _get_collision_slop,
-        _set_collision_slop,
-        doc="""Amount of overlap between shapes that is allowed.
+    @property
+    def collision_slop(self) -> float:
+        """Amount of overlap between shapes that is allowed.
 
         To improve stability, set this as high as you can without noticeable
         overlapping. It defaults to 0.1.
-        """,
-    )
+        """
+        return lib.cpSpaceGetCollisionSlop(self._space)
 
-    def _set_collision_bias(self, collision_bias: float) -> None:
-        cp.cpSpaceSetCollisionBias(self._space, collision_bias)
+    @collision_slop.setter
+    def collision_slop(self, collision_slop: float) -> None:
+        lib.cpSpaceSetCollisionSlop(self._space, collision_slop)
 
-    def _get_collision_bias(self) -> float:
-        return cp.cpSpaceGetCollisionBias(self._space)
-
-    collision_bias = property(
-        _get_collision_bias,
-        _set_collision_bias,
-        doc="""Determines how fast overlapping shapes are pushed apart.
+    @property
+    def collision_bias(self) -> float:
+        """Determines how fast overlapping shapes are pushed apart.
 
         Pymunk allows fast moving objects to overlap, then fixes the overlap
         over time. Overlapping objects are unavoidable even if swept
@@ -322,19 +314,16 @@ class Space(PickleMixin, object):
 
         ..Note::
             Very very few games will need to change this value.
-        """,
-    )
+        """
+        return lib.cpSpaceGetCollisionBias(self._space)
 
-    def _set_collision_persistence(self, collision_persistence: float) -> None:
-        cp.cpSpaceSetCollisionPersistence(self._space, collision_persistence)
+    @collision_bias.setter
+    def collision_bias(self, collision_bias: float) -> None:
+        lib.cpSpaceSetCollisionBias(self._space, collision_bias)
 
-    def _get_collision_persistence(self) -> float:
-        return cp.cpSpaceGetCollisionPersistence(self._space)
-
-    collision_persistence = property(
-        _get_collision_persistence,
-        _set_collision_persistence,
-        doc="""The number of frames the space keeps collision solutions
+    @property
+    def collision_persistence(self) -> float:
+        """The number of frames the space keeps collision solutions
         around for.
 
         Helps prevent jittering contacts from getting worse. This defaults
@@ -342,19 +331,20 @@ class Space(PickleMixin, object):
 
         ..Note::
             Very very few games will need to change this value.
-        """,
-    )
+        """
+        return lib.cpSpaceGetCollisionPersistence(self._space)
 
-    def _get_current_time_step(self) -> float:
-        return cp.cpSpaceGetCurrentTimeStep(self._space)
+    @collision_persistence.setter
+    def collision_persistence(self, collision_persistence: float) -> None:
+        lib.cpSpaceSetCollisionPersistence(self._space, collision_persistence)
 
-    current_time_step = property(
-        _get_current_time_step,
-        doc="""Retrieves the current (if you are in a callback from
+    @property
+    def current_time_step(self) -> float:
+        """Retrieves the current (if you are in a callback from
         Space.step()) or most recent (outside of a Space.step() call)
         timestep.
-        """,
-    )
+        """
+        return lib.cpSpaceGetCurrentTimeStep(self._space)
 
     def add(self, *objs: _AddableObjects) -> None:
         """Add one or many shapes, bodies or constraints (joints) to the space
@@ -412,8 +402,7 @@ class Space(PickleMixin, object):
 
     def _add_shape(self, shape: "Shape") -> None:
         """Adds a shape to the space"""
-        # print("addshape", self._space, shape)
-        assert shape._id not in self._shapes, "Shape already added to space."
+        assert shape not in self._shapes, "Shape already added to space."
         assert (
             shape.space == None
         ), "Shape already added to another space. A shape can only be in one space at a time."
@@ -422,43 +411,51 @@ class Space(PickleMixin, object):
             shape.body.space == self
         ), "The shape's body must be added to the space before (or at the same time) as the shape."
 
-        shape._space = weakref.proxy(self)
-        self._shapes[shape._id] = shape
-        cp.cpSpaceAddShape(self._space, shape._shape)
+        shape._space = weakref.ref(self)
+        self._shapes[shape] = None
+        lib.cpSpaceAddShape(self._space, shape._shape)
 
     def _add_body(self, body: "Body") -> None:
         """Adds a body to the space"""
         assert body not in self._bodies, "Body already added to this space."
         assert body.space == None, "Body already added to another space."
 
-        body._space = weakref.proxy(self)
+        body._space = weakref.ref(self)
         self._bodies[body] = None
-        cp.cpSpaceAddBody(self._space, body._body)
+        self._bodies_to_check.add(body)
+        lib.cpSpaceAddBody(self._space, body._body)
 
     def _add_constraint(self, constraint: "Constraint") -> None:
         """Adds a constraint to the space"""
         assert constraint not in self._constraints, "Constraint already added to space."
 
+        assert (
+            constraint.a.body_type == Body.DYNAMIC
+            or constraint.b.body_type == Body.DYNAMIC
+        ), "At leasts one of a constraint's bodies must be DYNAMIC."
+
         self._constraints[constraint] = None
-        cp.cpSpaceAddConstraint(self._space, constraint._constraint)
+        lib.cpSpaceAddConstraint(self._space, constraint._constraint)
 
     def _remove_shape(self, shape: "Shape") -> None:
         """Removes a shape from the space"""
-        assert shape._id in self._shapes, "shape not in space, already removed?"
-        self._removed_shapes[shape._id] = shape
-        shape._space = None
+        assert shape in self._shapes, "shape not in space, already removed?"
+        self._removed_shapes[shape] = None
+        shape._space = _dead_ref
         # During GC at program exit sometimes the shape might already be removed. Then skip this step.
-        if cp.cpSpaceContainsShape(self._space, shape._shape):
-            cp.cpSpaceRemoveShape(self._space, shape._shape)
-        del self._shapes[shape._id]
+        if lib.cpSpaceContainsShape(self._space, shape._shape):
+            lib.cpSpaceRemoveShape(self._space, shape._shape)
+        del self._shapes[shape]
 
     def _remove_body(self, body: "Body") -> None:
         """Removes a body from the space"""
         assert body in self._bodies, "body not in space, already removed?"
-        body._space = None
+        body._space = _dead_ref
+        if body in self._bodies_to_check:
+            self._bodies_to_check.remove(body)
         # During GC at program exit sometimes the shape might already be removed. Then skip this step.
-        if cp.cpSpaceContainsBody(self._space, body._body):
-            cp.cpSpaceRemoveBody(self._space, body._body)
+        if lib.cpSpaceContainsBody(self._space, body._body):
+            lib.cpSpaceRemoveBody(self._space, body._body)
         del self._bodies[body]
 
     def _remove_constraint(self, constraint: "Constraint") -> None:
@@ -468,47 +465,44 @@ class Space(PickleMixin, object):
         ), "constraint not in space, already removed?"
         # print("remove", constraint, constraint._constraint, self._constraints)
         # During GC at program exit sometimes the constraint might already be removed. Then skip this steip.
-        if cp.cpSpaceContainsConstraint(self._space, constraint._constraint):
-            cp.cpSpaceRemoveConstraint(self._space, constraint._constraint)
+        if lib.cpSpaceContainsConstraint(self._space, constraint._constraint):
+            lib.cpSpaceRemoveConstraint(self._space, constraint._constraint)
         del self._constraints[constraint]
 
     def reindex_shape(self, shape: Shape) -> None:
         """Update the collision detection data for a specific shape in the
         space.
         """
-        cp.cpSpaceReindexShape(self._space, shape._shape)
+        lib.cpSpaceReindexShape(self._space, shape._shape)
 
     def reindex_shapes_for_body(self, body: Body) -> None:
         """Reindex all the shapes for a certain body."""
-        cp.cpSpaceReindexShapesForBody(self._space, body._body)
+        lib.cpSpaceReindexShapesForBody(self._space, body._body)
 
     def reindex_static(self) -> None:
         """Update the collision detection info for the static shapes in the
         space. You only need to call this if you move one of the static shapes.
         """
-        cp.cpSpaceReindexStatic(self._space)
+        lib.cpSpaceReindexStatic(self._space)
 
-    def _get_threads(self) -> int:
+    @property
+    def threads(self) -> int:
+        """The number of threads to use for running the step function.
+
+        Only valid when the Space was created with threaded=True. Currently the
+        max limit is 2, setting a higher value wont have any effect. The
+        default is 1 regardless if the Space was created with threaded=True,
+        to keep determinism in the simulation. Note that Windows does not
+        support the threaded solver.
+        """
         if self.threaded:
-            return int(cp.cpHastySpaceGetThreads(self._space))
+            return int(lib.cpHastySpaceGetThreads(self._space))
         return 1
 
-    def _set_threads(self, n: int) -> None:
+    @threads.setter
+    def threads(self, n: int) -> None:
         if self.threaded:
-            cp.cpHastySpaceSetThreads(self._space, n)
-
-    threads = property(
-        _get_threads,
-        _set_threads,
-        doc="""The number of threads to use for running the step function. 
-        
-        Only valid when the Space was created with threaded=True. Currently the 
-        max limit is 2, setting a higher value wont have any effect. The 
-        default is 1 regardless if the Space was created with threaded=True, 
-        to keep determinism in the simulation. Note that Windows does not 
-        support the threaded solver.
-        """,
-    )
+            lib.cpHastySpaceSetThreads(self._space, n)
 
     def use_spatial_hash(self, dim: float, count: int) -> None:
         """Switch the space to use a spatial hash instead of the bounding box
@@ -532,7 +526,7 @@ class Space(PickleMixin, object):
         the shape to be inserted into many cells, setting it too low will
         cause too many objects into the same hash slot.
 
-        count is the suggested minimum number of cells in the hash table. If
+        count is the minimum number of cells in the hash table. If
         there are too few cells, the spatial hash will return many false
         positives. Too many cells will be hard on the cache and waste memory.
         Setting count to ~10x the number of objects in the space is probably a
@@ -541,7 +535,7 @@ class Space(PickleMixin, object):
         :param dim: the size of the hash cells
         :param count: the suggested minimum number of cells in the hash table
         """
-        cp.cpSpaceUseSpatialHash(self._space, dim, count)
+        lib.cpSpaceUseSpatialHash(self._space, dim, count)
 
     def step(self, dt: float) -> None:
         """Update the space for the given time step.
@@ -565,13 +559,20 @@ class Space(PickleMixin, object):
 
         :param dt: Time step length
         """
+
+        for b in self._bodies_to_check:
+            assert b.body_type != Body.DYNAMIC or (
+                b.mass > 0 and b.mass < math.inf and b.moment > 0
+            ), f"Dynamic bodies must have a mass > 0 and < inf and moment > 0. {b} has mass {b.mass}, moment {b.moment}."
+        self._bodies_to_check.clear()
+
         try:
             self._locked = True
             if self.threaded:
-                cp.cpHastySpaceStep(self._space, dt)
+                lib.cpHastySpaceStep(self._space, dt)
             else:
-                cp.cpSpaceStep(self._space, dt)
-            self._removed_shapes = {}
+                lib.cpSpaceStep(self._space, dt)
+            self._removed_shapes.clear()
         finally:
             self._locked = False
         self.add(*self._add_later)
@@ -583,92 +584,117 @@ class Space(PickleMixin, object):
         for key in self._post_step_callbacks:
             self._post_step_callbacks[key](self)
 
-        self._post_step_callbacks = {}
+        self._post_step_callbacks.clear()
 
-    def add_collision_handler(
-        self, collision_type_a: int, collision_type_b: int
-    ) -> CollisionHandler:
-        """Return the :py:class:`CollisionHandler` for collisions between
-        objects of type collision_type_a and collision_type_b.
-
-        Fill the desired collision callback functions, for details see the
-        :py:class:`CollisionHandler` object.
+    def on_collision(
+        self,
+        collision_type_a: Optional[int] = None,
+        collision_type_b: Optional[int] = None,
+        begin: Optional[_CollisionCallback] = None,
+        pre_solve: Optional[_CollisionCallback] = None,
+        post_solve: Optional[_CollisionCallback] = None,
+        separate: Optional[_CollisionCallback] = None,
+        data: Any = None,
+    ) -> None:
+        """Set callbacks that will be called during the 4 phases of collision
+        handling.
 
         Whenever shapes with collision types (:py:attr:`Shape.collision_type`)
-        a and b collide, this handler will be used to process the collision
-        events. When a new collision handler is created, the callbacks will all be
-        set to builtin callbacks that perform the default behavior (call the
-        wildcard handlers, and accept all collisions).
+        a and b collide, the callback will be used to process the collision
+        events. If no callback is set, the default is to process collision
+        normally.
 
-        :param int collision_type_a: Collision type a
-        :param int collision_type_b: Collision type b
+        Its possible to pass in None for one or both of the collision types.
+        None matches any collision type on a Shape. However, if
+        collision_type_a is None, then collision_type_b must also be None.
 
-        :rtype: :py:class:`CollisionHandler`
+        If you call this multiple times with the same combination of
+        collision_type_a and collision_type_b, then the last call will
+        overwrite the earlier.
+
+        If multiple callbacks match the collision, the order will be that the
+        most specific handler is called first.
+
+        Callback phases:
+
+            - **begin**: Two shapes just started touching for the first time
+              this step.
+            - **pre_solve**: Two shapes are touching during this step, before
+              collision resolution. You may override collision values using
+              Arbiter.friction, Arbiter.elasticity or Arbiter.surfaceVelocity
+              to provide custom friction, elasticity, or surface velocity
+              values. See Arbiter for more info.
+            - **post_solve**: Two shapes are touching and their collision response
+              has been processed. You can retrieve the collision impulse or
+              kinetic energy at this time if you want to use it to calculate
+              sound volumes or damage amounts. See Arbiter for more info.
+            - **separate**: Two shapes have just stopped touching for the
+              first time this step. To ensure that begin()/separate() are
+              always called in balanced pairs, it will also be called when
+              removing a shape while its in contact with something or when
+              de-allocating the space.
+
+        From each callback you can set process_collision on the Arbiter,
+        which decides if the collision should be processed or not.
+
+        data will be passed in to the callback function unchanged.
+
         """
-        key = min(collision_type_a, collision_type_b), max(
-            collision_type_a, collision_type_b
-        )
-        if key in self._handlers:
-            return self._handlers[key]
+        # key = min(collision_type_a, collision_type_b), max(
+        #     collision_type_a, collision_type_b
+        # )
+        assert (
+            collision_type_a != None or collision_type_b == None
+        ), "collision_type_a can not be None if collision_type_b is not None. Please swap them."
 
-        h = cp.cpSpaceAddCollisionHandler(
-            self._space, collision_type_a, collision_type_b
-        )
-        ch = CollisionHandler(h, self)
-        self._handlers[key] = ch
-        return ch
+        key = collision_type_a, collision_type_b
+        if key not in self._handlers:
+            # CP_WILDCARD_COLLISION_TYPE
+            wildcard = int(ffi.cast("uintptr_t", ~0))
+            if collision_type_a == None:
+                collision_type_a = wildcard
 
-    def add_wildcard_collision_handler(self, collision_type_a: int) -> CollisionHandler:
-        """Add a wildcard collision handler for given collision type.
+            if collision_type_b == None:
+                collision_type_b = wildcard
 
-        This handler will be used any time an object with this type collides
-        with another object, regardless of its type. A good example is a
-        projectile that should be destroyed the first time it hits anything.
-        There may be a specific collision handler and two wildcard handlers.
-        It's up to the specific handler to decide if and when to call the
-        wildcard handlers and what to do with their return values.
+            h = lib.cpSpaceAddCollisionHandler(
+                self._space, collision_type_a, collision_type_b
+            )
+            ch = CollisionHandler(h, self)
+            self._handlers[key] = ch
+        else:
+            ch = self._handlers[key]
 
-        When a new wildcard handler is created, the callbacks will all be
-        set to builtin callbacks that perform the default behavior. (accept
-        all collisions in :py:func:`~CollisionHandler.begin` and
-        :py:func:`~CollisionHandler.pre_solve`, or do nothing for
-        :py:func:`~CollisionHandler.post_solve` and
-        :py:func:`~CollisionHandler.separate`.
+        # to avoid circular dep
+        from . import empty_callback
 
-        :param int collision_type_a: Collision type
-        :rtype: :py:class:`CollisionHandler`
-        """
-
-        if collision_type_a in self._handlers:
-            return self._handlers[collision_type_a]
-
-        h = cp.cpSpaceAddWildcardHandler(self._space, collision_type_a)
-        ch = CollisionHandler(h, self)
-        self._handlers[collision_type_a] = ch
-        return ch
-
-    def add_default_collision_handler(self) -> CollisionHandler:
-        """Return a reference to the default collision handler or that is
-        used to process all collisions that don't have a more specific
-        handler.
-
-        The default behavior for each of the callbacks is to call
-        the wildcard handlers, ANDing their return values together if
-        applicable.
-        """
-        if None in self._handlers:
-            return self._handlers[None]
-
-        _h = cp.cpSpaceAddDefaultCollisionHandler(self._space)
-        h = CollisionHandler(_h, self)
-        self._handlers[None] = h
-        return h
+        if begin == empty_callback:
+            ch.begin = None
+        elif begin != None:
+            ch.begin = begin
+            ch.data["begin"] = data
+        if pre_solve == empty_callback:
+            ch.pre_solve = None
+        elif pre_solve != None:
+            ch.pre_solve = pre_solve
+            ch.data["pre_solve"] = data
+        if post_solve == empty_callback:
+            ch.post_solve = None
+        elif post_solve != None:
+            ch.post_solve = post_solve
+            ch.data["post_solve"] = data
+        if separate == empty_callback:
+            ch.separate = None
+        elif separate != None:
+            ch.separate = separate
+            ch.data["separate"] = data
+        return
 
     def add_post_step_callback(
         self,
         callback_function: Callable[
             ..., None
-        ],  # TODO: Fix me once PEP-612 is implemented
+        ],  # TODO: Fix me once PEP-612 is implemented (py 3.10)
         key: Hashable,
         *args: Any,
         **kwargs: Any,
@@ -712,8 +738,8 @@ class Space(PickleMixin, object):
         return True
 
     def point_query(
-        self, point: Tuple[float, float], max_distance: float, shape_filter: ShapeFilter
-    ) -> List[PointQueryInfo]:
+        self, point: tuple[float, float], max_distance: float, shape_filter: ShapeFilter
+    ) -> list[PointQueryInfo]:
         """Query space at point for shapes within the given distance range.
 
         The filter is applied to the query and follows the same rules as the
@@ -722,12 +748,10 @@ class Space(PickleMixin, object):
         the point must be a under a certain depth within a shape to be
         considered a match.
 
+        Sensor shapes are included in the result.
+
         See :py:class:`ShapeFilter` for details about how the shape_filter
         parameter can be used.
-
-        .. Note::
-            Sensor shapes are included in the result (In
-            :py:meth:`Space.point_query_nearest` they are not)
 
         :param point: Where to check for collision in the Space
         :type point: :py:class:`~vec2d.Vec2d` or (float,float)
@@ -737,35 +761,21 @@ class Space(PickleMixin, object):
         :rtype: [:py:class:`PointQueryInfo`]
         """
         assert len(point) == 2
-        query_hits: List[PointQueryInfo] = []
+        query_hits: list[PointQueryInfo] = []
         d = (self, query_hits)
         data = ffi.new_handle(d)
-        cp.cpSpacePointQuery(
+        lib.cpSpacePointQuery(
             self._space,
             point,
             max_distance,
             shape_filter,
-            cp.ext_cpSpacePointQueryFunc,
+            lib.ext_cpSpacePointQueryFunc,
             data,
         )
         return query_hits
 
-    def _get_shape(self, _shape: Any) -> Optional[Shape]:
-        if not bool(_shape):
-            return None
-
-        shapeid = int(ffi.cast("int", cp.cpShapeGetUserData(_shape)))
-        # return self._shapes[hashid_private]
-
-        if shapeid in self._shapes:
-            return self._shapes[shapeid]
-        elif shapeid in self._removed_shapes:
-            return self._removed_shapes[shapeid]
-        else:
-            return None
-
     def point_query_nearest(
-        self, point: Tuple[float, float], max_distance: float, shape_filter: ShapeFilter
+        self, point: tuple[float, float], max_distance: float, shape_filter: ShapeFilter
     ) -> Optional[PointQueryInfo]:
         """Query space at point the nearest shape within the given distance
         range.
@@ -780,8 +790,7 @@ class Space(PickleMixin, object):
         parameter can be used.
 
         .. Note::
-            Sensor shapes are not included in the result (In
-            :py:meth:`Space.point_query` they are)
+            Sensor shapes are included in the result. (Changed in Pymunk 7.0)
 
         :param point: Where to check for collision in the Space
         :type point: :py:class:`~vec2d.Vec2d` or (float,float)
@@ -792,11 +801,11 @@ class Space(PickleMixin, object):
         """
         assert len(point) == 2
         info = ffi.new("cpPointQueryInfo *")
-        _shape = cp.cpSpacePointQueryNearest(
+        _shape = lib.cpSpacePointQueryNearest(
             self._space, point, max_distance, shape_filter, info
         )
 
-        shape = self._get_shape(_shape)
+        shape = Shape._from_cp_shape(_shape)
 
         if shape != None:
             return PointQueryInfo(
@@ -809,11 +818,11 @@ class Space(PickleMixin, object):
 
     def segment_query(
         self,
-        start: Tuple[float, float],
-        end: Tuple[float, float],
+        start: tuple[float, float],
+        end: tuple[float, float],
         radius: float,
         shape_filter: ShapeFilter,
-    ) -> List[SegmentQueryInfo]:
+    ) -> list[SegmentQueryInfo]:
         """Query space along the line segment from start to end with the
         given radius.
 
@@ -823,9 +832,7 @@ class Space(PickleMixin, object):
         See :py:class:`ShapeFilter` for details about how the shape_filter
         parameter can be used.
 
-        .. Note::
-            Sensor shapes are included in the result (In
-            :py:meth:`Space.segment_query_first` they are not)
+        Sensor shapes are included in the result.
 
         :param start: Starting point
         :param end: End point
@@ -836,26 +843,26 @@ class Space(PickleMixin, object):
         """
         assert len(start) == 2
         assert len(end) == 2
-        query_hits: List[SegmentQueryInfo] = []
+        query_hits: list[SegmentQueryInfo] = []
 
         d = (self, query_hits)
         data = ffi.new_handle(d)
 
-        cp.cpSpaceSegmentQuery(
+        lib.cpSpaceSegmentQuery(
             self._space,
             start,
             end,
             radius,
             shape_filter,
-            cp.ext_cpSpaceSegmentQueryFunc,
+            lib.ext_cpSpaceSegmentQueryFunc,
             data,
         )
         return query_hits
 
     def segment_query_first(
         self,
-        start: Tuple[float, float],
-        end: Tuple[float, float],
+        start: tuple[float, float],
+        end: tuple[float, float],
         radius: float,
         shape_filter: ShapeFilter,
     ) -> Optional[SegmentQueryInfo]:
@@ -866,8 +873,7 @@ class Space(PickleMixin, object):
         collision detection.
 
         .. Note::
-            Sensor shapes are not included in the result (In
-            :py:meth:`Space.segment_query` they are)
+            Sensor shapes are included in the result. (Changed in Pymunk 7.0)
 
         See :py:class:`ShapeFilter` for details about how the shape_filter
         parameter can be used.
@@ -877,11 +883,11 @@ class Space(PickleMixin, object):
         assert len(start) == 2
         assert len(end) == 2
         info = ffi.new("cpSegmentQueryInfo *")
-        _shape = cp.cpSpaceSegmentQueryFirst(
+        _shape = lib.cpSpaceSegmentQueryFirst(
             self._space, start, end, radius, shape_filter, info
         )
 
-        shape = self._get_shape(_shape)
+        shape = Shape._from_cp_shape(_shape)
         if shape != None:
             return SegmentQueryInfo(
                 shape,
@@ -891,14 +897,13 @@ class Space(PickleMixin, object):
             )
         return None
 
-    def bb_query(self, bb: "BB", shape_filter: ShapeFilter) -> List[Shape]:
+    def bb_query(self, bb: "BB", shape_filter: ShapeFilter) -> list[Shape]:
         """Query space to find all shapes near bb.
 
         The filter is applied to the query and follows the same rules as the
         collision detection.
 
-        .. Note::
-            Sensor shapes are included in the result
+        Sensor shapes are included in the result
 
         :param bb: Bounding box
         :param shape_filter: Shape filter
@@ -906,21 +911,20 @@ class Space(PickleMixin, object):
         :rtype: [:py:class:`Shape`]
         """
 
-        query_hits: List[Shape] = []
+        query_hits: list[Shape] = []
 
         d = (self, query_hits)
         data = ffi.new_handle(d)
 
-        cp.cpSpaceBBQuery(
-            self._space, bb, shape_filter, cp.ext_cpSpaceBBQueryFunc, data
+        lib.cpSpaceBBQuery(
+            self._space, bb, shape_filter, lib.ext_cpSpaceBBQueryFunc, data
         )
         return query_hits
 
-    def shape_query(self, shape: Shape) -> List[ShapeQueryInfo]:
+    def shape_query(self, shape: Shape) -> list[ShapeQueryInfo]:
         """Query a space for any shapes overlapping the given shape
 
-        .. Note::
-            Sensor shapes are included in the result
+        Sensor shapes are included in the result
 
         :param shape: Shape to query with
         :type shape: :py:class:`Circle`, :py:class:`Poly` or :py:class:`Segment`
@@ -928,12 +932,12 @@ class Space(PickleMixin, object):
         :rtype: [:py:class:`ShapeQueryInfo`]
         """
 
-        query_hits: List[ShapeQueryInfo] = []
+        query_hits: list[ShapeQueryInfo] = []
         d = (self, query_hits)
         data = ffi.new_handle(d)
 
-        cp.cpSpaceShapeQuery(
-            self._space, shape._shape, cp.ext_cpSpaceShapeQueryFunc, data
+        lib.cpSpaceShapeQuery(
+            self._space, shape._shape, lib.ext_cpSpaceShapeQueryFunc, data
         )
 
         return query_hits
@@ -962,7 +966,7 @@ class Space(PickleMixin, object):
             options._options.data = h
 
             with options:
-                cp.cpSpaceDebugDraw(self._space, options._options)
+                lib.cpSpaceDebugDraw(self._space, options._options)
         else:
             for shape in self.shapes:
                 options.draw_shape(shape)
@@ -976,10 +980,10 @@ class Space(PickleMixin, object):
     #     """
     #     pass
 
-    def _get_arbiters(self) -> List[ffi.CData]:
-        _arbiters: List[ffi.CData] = []
+    def _get_arbiters(self) -> list[ffi.CData]:
+        _arbiters: list[ffi.CData] = []
         data = ffi.new_handle(_arbiters)
-        cp.cpSpaceEachCachedArbiter(self._space, cp.ext_cpArbiterIteratorFunc, data)
+        lib.cpSpaceEachCachedArbiter(self._space, lib.ext_cpArbiterIteratorFunc, data)
         return _arbiters
 
     def __getstate__(self) -> _State:
@@ -992,35 +996,38 @@ class Space(PickleMixin, object):
 
         d["special"].append(("pymunk_version", _version.version))
         # bodies needs to be added to the state before their shapes.
-        d["special"].append(("bodies", self.bodies))
+        d["special"].append(("bodies", list(self.bodies)))
         if self._static_body != None:
             # print("getstate", self._static_body)
             d["special"].append(("_static_body", self._static_body))
 
-        d["special"].append(("shapes", self.shapes))
-        d["special"].append(("constraints", self.constraints))
+        d["special"].append(("shapes", list(self.shapes)))
+        d["special"].append(("constraints", list(self.constraints)))
+
+        # to avoid circular dep
+        from . import empty_callback
 
         handlers = []
         for k, v in self._handlers.items():
-            h: Dict[str, Any] = {}
-            if v._begin is not None:
+            h: dict[str, Any] = {}
+            if v._begin != empty_callback:
                 h["_begin"] = v._begin
-            if v._pre_solve is not None:
+            if v._pre_solve != empty_callback:
                 h["_pre_solve"] = v._pre_solve
-            if v._post_solve is not None:
+            if v._post_solve != empty_callback:
                 h["_post_solve"] = v._post_solve
-            if v._separate is not None:
+            if v._separate != empty_callback:
                 h["_separate"] = v._separate
             handlers.append((k, h))
 
         d["special"].append(("_handlers", handlers))
 
         d["special"].append(
-            ("shapeIDCounter", cp.cpSpaceGetShapeIDCounter(self._space))
+            ("shapeIDCounter", lib.cpSpaceGetShapeIDCounter(self._space))
         )
-        d["special"].append(("stamp", cp.cpSpaceGetTimestamp(self._space)))
+        d["special"].append(("stamp", lib.cpSpaceGetTimestamp(self._space)))
         d["special"].append(
-            ("currentTimeStep", cp.cpSpaceGetCurrentTimeStep(self._space))
+            ("currentTimeStep", lib.cpSpaceGetCurrentTimeStep(self._space))
         )
 
         _arbs = self._get_arbiters()
@@ -1045,14 +1052,14 @@ class Space(PickleMixin, object):
             elif k == "bodies":
                 self.add(*v)
             elif k == "_static_body":
-                # _ = cp.cpSpaceSetStaticBody(self._space, v._body)
+                # _ = lib.cpSpaceSetStaticBody(self._space, v._body)
                 # v._space = self
                 # self._static_body = v
                 # print("setstate", v, self._static_body)
                 self._static_body = v
                 self._setup_static_body(v)
                 # self._static_body._space = weakref.proxy(self)
-                # cp.cpSpaceAddBody(self._space, v._body)
+                # lib.cpSpaceAddBody(self._space, v._body)
                 # self.add(v)
 
             elif k == "shapes":
@@ -1062,28 +1069,51 @@ class Space(PickleMixin, object):
                 self.add(*v)
             elif k == "_handlers":
                 for k2, hd in v:
-                    if k2 == None:
-                        h = self.add_default_collision_handler()
-                    elif isinstance(k2, tuple):
-                        h = self.add_collision_handler(k2[0], k2[1])
-                    else:
-                        h = self.add_wildcard_collision_handler(k2)
+                    begin = pre_solve = post_solve = separate = None
                     if "_begin" in hd:
-                        h.begin = hd["_begin"]
+                        begin = hd["_begin"]
                     if "_pre_solve" in hd:
-                        h.pre_solve = hd["_pre_solve"]
+                        pre_solve = hd["_pre_solve"]
                     if "_post_solve" in hd:
-                        h.post_solve = hd["_post_solve"]
+                        post_solve = hd["_post_solve"]
                     if "_separate" in hd:
-                        h.separate = hd["_separate"]
+                        separate = hd["_separate"]
+                    if k2 == None:
+                        self.on_collision(
+                            None,
+                            None,
+                            begin=begin,
+                            pre_solve=pre_solve,
+                            post_solve=post_solve,
+                            separate=separate,
+                        )
+                    elif isinstance(k2, tuple):
+                        self.on_collision(
+                            k2[0],
+                            k2[1],
+                            begin=begin,
+                            pre_solve=pre_solve,
+                            post_solve=post_solve,
+                            separate=separate,
+                        )
+                    else:
+                        self.on_collision(
+                            k2,
+                            None,
+                            begin=begin,
+                            pre_solve=pre_solve,
+                            post_solve=post_solve,
+                            separate=separate,
+                        )
+
             elif k == "stamp":
-                cp.cpSpaceSetTimestamp(self._space, v)
+                lib.cpSpaceSetTimestamp(self._space, v)
             elif k == "shapeIDCounter":
-                cp.cpSpaceSetShapeIDCounter(self._space, v)
+                lib.cpSpaceSetShapeIDCounter(self._space, v)
             elif k == "currentTimeStep":
-                cp.cpSpaceSetCurrentTimeStep(self._space, v)
+                lib.cpSpaceSetCurrentTimeStep(self._space, v)
             elif k == "arbiters":
                 for d in v:
-                    # cp.cpSpaceTest(self._space)
+                    # lib.cpSpaceTest(self._space)
                     _arbiter = _arbiter_from_dict(d, self)
-                    cp.cpSpaceAddCachedArbiter(self._space, _arbiter)
+                    lib.cpSpaceAddCachedArbiter(self._space, _arbiter)
