@@ -1,7 +1,10 @@
+import gc
 import unittest
-from typing import Any, List
+import weakref
+from typing import Any
 
 import pymunk as p
+from pymunk._weakkeysview import WeakKeysView
 from pymunk.vec2d import Vec2d
 
 
@@ -42,7 +45,7 @@ class UnitTestBugs(unittest.TestCase):
 
         _logger = logging.getLogger(__name__)
 
-        def make() -> List[Any]:
+        def make() -> list[Any]:
             s = p.Space()
             b1 = p.Body(1, 2)
             c1 = p.Circle(b1, 2)
@@ -123,7 +126,7 @@ class UnitTestBugs(unittest.TestCase):
             # space.add_post_step_callback(space.remove, first_shape, first_shape.body)
             # space.remove(c1)
 
-        space.add_collision_handler(2, 0).separate = remove_first
+        space.on_collision(2, 0, separate=remove_first)
         # print(1)
         space.step(1.0 / 60)
         # print(2)
@@ -134,14 +137,14 @@ class UnitTestBugs(unittest.TestCase):
     def testX(self) -> None:
         space = p.Space()
 
-        b1 = p.Body()
+        b1 = p.Body(1, 1)
         c1 = p.Circle(b1, 10)
         c1.collision_type = 2
 
-        b2 = p.Body()
+        b2 = p.Body(1, 2)
         c2 = p.Circle(b2, 10)
 
-        b3 = p.Body()
+        b3 = p.Body(1, 3)
         c3 = p.Circle(b3, 10)
 
         # b1.position = 0, 0
@@ -163,10 +166,38 @@ class UnitTestBugs(unittest.TestCase):
             # space.add_post_step_callback(space.remove, first_shape, first_shape.body)
             # space.remove(c1)
 
-        space.add_collision_handler(2, 0).separate = separate
+        space.on_collision(2, 0, separate=separate)
         # print(1)
         space.step(1)
         # print(2)
         b2.position = 22, 0
         space.step(1)
         # print(3)
+
+    def testWeakKeysView(self) -> None:
+        x1, x2, x3 = p.Body(1), p.Body(2), p.Body(3)
+
+        d: weakref.WeakKeyDictionary[p.Body, int] = weakref.WeakKeyDictionary()
+
+        d[x1] = 1
+        d[x2] = 2
+
+        keys1 = WeakKeysView(d)
+        keys2 = WeakKeysView(d)
+
+        del x2
+        d[x3] = 3
+
+        iterations = 0
+        for x in keys2:
+            iterations += 1
+            del x3
+            gc.collect()
+
+        self.assertEqual(iterations, 1)
+        gc.collect()
+
+        self.assertEqual(len(d), 1)
+        self.assertEqual(list(keys1), [x1])
+        self.assertEqual(list(keys2), [x1])
+        self.assertEqual(list(d.keys()), [x1])
