@@ -8,10 +8,9 @@ s.add(pymunk.Circle(s.static_body, 5))
 b = pymunk.Body(1,10)
 c = pymunk.Circle(b, 5)
 s.add(b, c)
-h = s.add_default_collision_handler()
 def f(arb, s, data):
     return False
-h.pre_solve = f
+s.on_collision(pre_solve=f)
 """
 
 print(min(timeit.repeat("s.step(0.01)", setup=s, repeat=10)))
